@@ -35,7 +40,22 @@ classifiers = [
     'Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)',
 ]
 
-requires-python = ">=3.8"
+requires-python = ">=3.9"
+dependencies = [
+    "z3-solver>=4.8.0"
+]
+
+[project.optional-dependencies]
+test = ["pytest>=7.2", "mnt.pyfiction"]
+docs = [
+    "setuptools-scm>=8.1",
+    "sphinx==8.1.3; python_version >= '3.10'",
+    "breathe==4.35.0",
+    "sphinx-rtd-theme==3.0.2",
+    "sphinx-tabs==3.4.7",
+    "z3-solver>=4.8.0",
+    "mnt.pyfiction"
+]
 
 [project.urls]
 Source = 'https://github.com/cda-tum/fiction'
