@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 #
 # fiction documentation build configuration file
 #
@@ -18,6 +16,15 @@
 # sys.path.insert(0, os.path.abspath('.'))
 
 
+import os
+import subprocess
+import xml.etree.ElementTree as ET
+from typing import ClassVar
+
+from docutils import nodes
+from docutils.parsers.rst import Directive
+from sphinx.application import Sphinx
+
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
@@ -27,48 +34,48 @@
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.mathjax', 'sphinx.ext.viewcode', 'sphinx_tabs.tabs', 'sphinx.ext.autodoc', 'breathe']
+extensions = ["sphinx.ext.mathjax", "sphinx.ext.viewcode", "sphinx_tabs.tabs", "sphinx.ext.autodoc", "breathe"]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # General information about the project.
-project = 'fiction'
-copyright = '2018-2023: Marcel Walter. 2023-present: Chair for Design Automation, Technical University of Munich'
-author = 'Marcel Walter, Jan Drewniok, Simon Hofmann, Benjamin Hien'
+project = "fiction"
+copyright = "2018-2023: Marcel Walter. 2023-present: Chair for Design Automation, Technical University of Munich (TUM)"
+author = "Marcel Walter, Jan Drewniok, Simon Hofmann, Benjamin Hien, Willem Lambooy"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The short X.Y version.
-version = 'v0.6.5'
+version = "v0.6.8"
 # The full version, including alpha/beta/rc tags.
-release = 'v0.6.5'
+release = "v0.6.8"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = 'en'
+language = "en"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = False
@@ -78,20 +85,18 @@ todo_include_todos = False
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-html_theme_options = {
-    "collapse_navigation": False
-}
+html_theme_options = {"collapse_navigation": False}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
@@ -99,19 +104,19 @@ html_static_path = ['_static']
 # This is required for the alabaster theme
 # refs: http://alabaster.readthedocs.io/en/latest/installation.html#sidebars
 html_sidebars = {
-    '**': [
-        'about.html',
-        'navigation.html',
-        'relations.html',  # needs 'show_related': True theme option to display
-        'searchbox.html',
-        'donate.html',
+    "**": [
+        "about.html",
+        "navigation.html",
+        "relations.html",  # needs 'show_related': True theme option to display
+        "searchbox.html",
+        "donate.html",
     ]
 }
 
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'fictiondoc'
+htmlhelp_basename = "fictiondoc"
 html_logo = "_static/mnt_sidebar.svg"
 
 # -- Options for LaTeX output ---------------------------------------------
@@ -120,15 +125,12 @@ latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
     # 'papersize': 'letterpaper',
-
     # The font size ('10pt', '11pt' or '12pt').
     #
     # 'pointsize': '10pt',
-
     # Additional stuff for the LaTeX preamble.
     #
     # 'preamble': '',
-
     # Latex figure (float) alignment
     #
     # 'figure_align': 'htbp',
@@ -138,18 +140,20 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'fiction.tex', 'fiction Documentation',
-     'Chair for Design Automation, Technical University of Munich', 'manual'),
+    (
+        master_doc,
+        "fiction.tex",
+        "fiction Documentation",
+        "Chair for Design Automation, Technical University of Munich",
+        "manual",
+    ),
 ]
 
 # -- Options for manual page output ---------------------------------------
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'fiction', 'fiction Documentation',
-     [author], 1)
-]
+man_pages = [(master_doc, "fiction", "fiction Documentation", [author], 1)]
 
 # -- Options for Texinfo output -------------------------------------------
 
@@ -157,38 +161,37 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'fiction', 'fiction Documentation',
-     author, 'fiction', 'Design Automation for Field-coupled Nanotechnologies.',
-     'Physical Design'),
+    (
+        master_doc,
+        "fiction",
+        "fiction Documentation",
+        author,
+        "fiction",
+        "Design Automation for Field-coupled Nanotechnologies.",
+        "Physical Design",
+    ),
 ]
 
 # -- Options for breathe --------------------------------------------------
 
-import os
-import subprocess
-
-read_the_docs_build = os.environ.get('READTHEDOCS', None) == 'True'
+read_the_docs_build = os.environ.get("READTHEDOCS", None) == "True"
 
 if read_the_docs_build:
-    subprocess.call('doxygen Doxyfile', shell=True)
+    subprocess.call("doxygen Doxyfile", shell=True)
 
 breathe_projects = {"fiction": "doxyxml/xml"}
 breathe_default_project = "fiction"
 
 # -- Custom directives ----------------------------------------------------
 
-from docutils import nodes
-from docutils.parsers.rst import Directive
-import xml.etree.ElementTree as ET
-
 
 class DocOverviewTableDirective(Directive):
     has_content = True
     required_arguments = 1
-    option_spec = {"column": str}
+    option_spec: ClassVar = {"column": str}
 
-    def run(self):
-        doc = ET.parse("doxyxml/xml/{}.xml".format(self.arguments[0]))
+    def run(self) -> list:
+        doc = ET.parse(f"doxyxml/xml/{self.arguments[0]}.xml")
 
         table = nodes.table()
         tgroup = nodes.tgroup(cols=2)
@@ -197,32 +200,33 @@ class DocOverviewTableDirective(Directive):
         tgroup += nodes.colspec(colwidth=50)
 
         # header
-        colname = self.options.get('column', "Function")
-        tgroup += nodes.thead('',
-                              nodes.row('', *[nodes.entry('', nodes.line(text=c)) for c in [colname, "Description"]]))
+        colname = self.options.get("column", "Function")
+        tgroup += nodes.thead(
+            "", nodes.row("", *[nodes.entry("", nodes.line(text=c)) for c in [colname, "Description"]])
+        )
 
         # rows
         tbody = nodes.tbody()
         for target in self.content:
-            for elem in doc.findall("./compounddef/sectiondef/memberdef/[name='%s']" % target):
-                ref = nodes.reference('', target, internal=True)
-                ref['refuri'] = '#{}'.format(elem.attrib["id"])
+            for elem in doc.findall(f"./compounddef/sectiondef/memberdef/[name='{target}']"):
+                ref = nodes.reference("", target, internal=True)
+                ref["refuri"] = "#{}".format(elem.attrib["id"])
 
                 reft = nodes.paragraph()
                 reft.extend([ref])
 
-                func = nodes.entry('', reft)
-                desc = nodes.entry('', nodes.line(text=elem.findtext("./briefdescription/para")))
+                func = nodes.entry("", reft)
+                desc = nodes.entry("", nodes.line(text=elem.findtext("./briefdescription/para")))
 
-                tbody += nodes.row('', func, desc)
+                tbody += nodes.row("", func, desc)
 
         tgroup += tbody
         table += tgroup
         return [table]
 
 
-def setup(app):
-    app.add_directive('doc_overview_table', DocOverviewTableDirective)
+def setup(app: Sphinx) -> None:
+    app.add_directive("doc_overview_table", DocOverviewTableDirective)
 
 
-autodoc_member_order = 'bysource'
+autodoc_member_order = "bysource"
