@@ -27,7 +27,7 @@
 namespace alice
 {
 /**
- *
+ * Executes *QuickSim* heuristic simulation for the current SiDB cell-level layout in store.
  */
 class quicksim_command : public command
 {
@@ -47,9 +47,10 @@ class quicksim_command : public command
                    true);
         add_option("--lambda_tf,-l", physical_params.lambda_tf, "Thomas-Fermi screening distance (unit: nm)", true);
         add_option("--mu_minus,-m", physical_params.mu_minus, "Energy transition level (0/-) (unit: eV)", true);
-        add_option("--iterations,-i", params.iteration_steps, "Number of iterations to run the simulation for", true);
-        add_option("--alpha,-a", params.alpha,
-                   "alpha parameter (should be reduced if not charge distribution can be determined)", true);
+        add_option("--iterations,-i", qs_params.iteration_steps, "Number of iterations to run the simulation for",
+                   true);
+        add_option("--alpha,-a", qs_params.alpha,
+                   "alpha parameter (should be reduced if no charge distribution can be determined)", true);
     }
 
   protected:
@@ -59,26 +60,24 @@ class quicksim_command : public command
     void execute() override
     {
         // reset sim result
-        sim_result_100      = {};
-        sim_result_111      = {};
-        min_energy          = std::numeric_limits<double>::infinity();
-        is_sidb_100_lattice = true;
+        sim_result = {};
+        min_energy = std::numeric_limits<double>::infinity();
 
         if (physical_params.epsilon_r <= 0)
         {
-            env->out() << "[e] epsilon_r must be positive" << std::endl;
+            env->out() << "[e] epsilon_r must be positive\n";
             reset_params();
             return;
         }
         if (physical_params.lambda_tf <= 0)
         {
-            env->out() << "[e] lambda_tf must be positive" << std::endl;
+            env->out() << "[e] lambda_tf must be positive\n";
             reset_params();
             return;
         }
-        if (params.alpha <= 0)
+        if (qs_params.alpha <= 0)
         {
-            env->out() << "[e] alpha must be positive" << std::endl;
+            env->out() << "[e] alpha must be positive\n";
             reset_params();
             return;
         }
@@ -88,7 +87,7 @@ class quicksim_command : public command
         // error case: empty cell layout store
         if (s.empty())
         {
-            env->out() << "[w] no cell layout in store" << std::endl;
+            env->out() << "[w] no cell layout in store\n";
             reset_params();
             return;
         }
@@ -99,73 +98,70 @@ class quicksim_command : public command
         {
             using Lyt = typename std::decay_t<decltype(lyt_ptr)>::element_type;
 
-            if constexpr (fiction::has_sidb_technology_v<Lyt>)
+            if constexpr (!fiction::has_sidb_technology_v<Lyt>)
             {
-                if constexpr (fiction::is_charge_distribution_surface_v<Lyt>)
-                {
-                    env->out() << fmt::format(
-                                      "[w] {} already possesses a charge distribution; no simulation is conducted",
-                                      get_name(lyt_ptr))
-                               << std::endl;
-                }
-                else
-                {
-                    params.simulation_parameters = physical_params;
+                env->out() << fmt::format("[e] '{}' is not an SiDB layout\n", get_name(lyt_ptr));
+                return;
+            }
 
-                    if constexpr (fiction::is_sidb_lattice_100_v<Lyt>)
-                    {
-                        is_sidb_100_lattice = true;
-                        sim_result_100      = fiction::quicksim(*lyt_ptr, params);
-                    }
-                    else if constexpr (fiction::is_sidb_lattice_111_v<Lyt>)
-                    {
-                        is_sidb_100_lattice = false;
-                        sim_result_111      = fiction::quicksim(*lyt_ptr, params);
-                    }
+            if constexpr (fiction::is_charge_distribution_surface_v<Lyt>)
+            {
+                env->out() << fmt::format(
+                    "[w] '{}' already possesses a charge distribution; no simulation is conducted\n",
+                    get_name(lyt_ptr));
+                return;
+            }
 
-                    else
+            qs_params.simulation_parameters = physical_params;
+
+            // To aid the compiler
+            if constexpr (fiction::has_sidb_technology_v<Lyt> && !fiction::is_charge_distribution_surface_v<Lyt>)
+            {
+                sim_result = fiction::quicksim(*lyt_ptr, qs_params);
+
+                if constexpr (fiction::is_sidb_lattice_100_v<Lyt>)
+                {
+
+                    if (std::get<sim_result_100>(sim_result).charge_distributions.empty())
                     {
-                        env->out() << "[e] no valid lattice orientation" << std::endl;
+                        env->out() << fmt::format("[e] no stable charge distribution could be determined for '{}'\n",
+                                                  get_name(lyt_ptr));
                         return;
                     }
 
-                    if (sim_result_100.charge_distributions.empty() && sim_result_111.charge_distributions.empty())
-                    {
-                        env->out() << fmt::format("[e] no stable charge distribution could be determined for {}",
-                                                  get_name(lyt_ptr))
-                                   << std::endl;
-                    }
-                    else
-                    {
-                        if constexpr (fiction::is_sidb_lattice_100_v<Lyt>)
-                        {
-                            const auto min_energy_distr =
-                                fiction::minimum_energy_distribution(sim_result_100.charge_distributions.cbegin(),
-                                                                     sim_result_100.charge_distributions.cend());
+                    const auto min_energy_distr =
+                        minimum_energy_distribution(std::get<sim_result_100>(sim_result).charge_distributions.cbegin(),
+                                                    std::get<sim_result_100>(sim_result).charge_distributions.cend());
 
-                            min_energy = min_energy_distr->get_system_energy();
-                            store<fiction::cell_layout_t>().extend() =
-                                std::make_shared<fiction::cds_sidb_100_cell_clk_lyt>(*min_energy_distr);
-                        }
-                        else if constexpr (fiction::is_sidb_lattice_111_v<Lyt>)
-                        {
-                            const auto min_energy_distr =
-                                fiction::minimum_energy_distribution(sim_result_111.charge_distributions.cbegin(),
-                                                                     sim_result_111.charge_distributions.cend());
-
-                            min_energy = min_energy_distr->get_system_energy();
-                            store<fiction::cell_layout_t>().extend() =
-                                std::make_shared<fiction::cds_sidb_111_cell_clk_lyt>(*min_energy_distr);
-                        }
-                    }
+                    min_energy = min_energy_distr->get_system_energy();
+                    store<fiction::cell_layout_t>().extend() =
+                        std::make_shared<fiction::cds_sidb_100_cell_clk_lyt>(*min_energy_distr);
                 }
-            }
-            else
-            {
-                env->out() << fmt::format("[e] {} is not an SiDB layout", get_name(lyt_ptr)) << std::endl;
+                else if constexpr (fiction::is_sidb_lattice_111_v<Lyt>)
+                {
+                    if (std::get<sim_result_111>(sim_result).charge_distributions.empty())
+                    {
+                        env->out() << fmt::format("[e] no stable charge distribution could be determined for '{}'\n",
+                                                  get_name(lyt_ptr));
+                        return;
+                    }
+
+                    const auto min_energy_distr =
+                        minimum_energy_distribution(std::get<sim_result_111>(sim_result).charge_distributions.cbegin(),
+                                                    std::get<sim_result_111>(sim_result).charge_distributions.cend());
+
+                    min_energy = min_energy_distr->get_system_energy();
+                    store<fiction::cell_layout_t>().extend() =
+                        std::make_shared<fiction::cds_sidb_111_cell_clk_lyt>(*min_energy_distr);
+                }
+                else
+                {
+                    env->out() << "[e] no valid lattice orientation\n";
+                }
             }
         };
 
+        // Dispatch on the layout currently in the store.
         std::visit(quicksim, s.current());
 
         reset_params();
@@ -179,22 +175,23 @@ class quicksim_command : public command
     /**
      * QuickSim parameters.
      */
-    fiction::quicksim_params params{};
+    fiction::quicksim_params qs_params{};
     /**
-     * Simulation result for H-Si(100)-2x1 surface.
+     * Type alias for H-Si(100)-2x1 simulation result.
      */
-    fiction::sidb_simulation_result<fiction::sidb_100_cell_clk_lyt> sim_result_100{};
+    using sim_result_100 = fiction::sidb_simulation_result<fiction::sidb_100_cell_clk_lyt>;
     /**
-     * Simulation result for H-Si(111)-1x1 surface.
+     * Type alias for H-Si(111)-1x1 simulation result.
      */
-    fiction::sidb_simulation_result<fiction::sidb_111_cell_clk_lyt> sim_result_111{};
+    using sim_result_111 = fiction::sidb_simulation_result<fiction::sidb_111_cell_clk_lyt>;
+    /**
+     * Simulation result for either the H-Si(100)-2x1 or the H-Si(111)-1x1 surface.
+     */
+    std::variant<sim_result_100, sim_result_111> sim_result;
     /**
      * Minimum energy.
      */
     double min_energy{std::numeric_limits<double>::infinity()};
-
-    bool is_sidb_100_lattice = true;
-
     /**
      * Logs the resulting information in a log file.
      *
@@ -204,46 +201,54 @@ class quicksim_command : public command
     {
         try
         {
-            if (is_sidb_100_lattice)
+            if (std::holds_alternative<sim_result_100>(sim_result))
             {
+                const auto& sim_res = std::get<sim_result_100>(sim_result);
+
                 return nlohmann::json{
-                    {"Algorithm name", sim_result_100.algorithm_name},
-                    {"Simulation runtime", sim_result_100.simulation_runtime.count()},
+                    {"Algorithm name", sim_res.algorithm_name},
+                    {"Simulation runtime", sim_res.simulation_runtime.count()},
                     {"Physical parameters",
-                     {{"epsilon_r", sim_result_100.simulation_parameters.epsilon_r},
-                      {"lambda_tf", sim_result_100.simulation_parameters.lambda_tf},
-                      {"mu_minus", sim_result_100.simulation_parameters.mu_minus}}},
+                     {{"epsilon_r", sim_res.simulation_parameters.epsilon_r},
+                      {"lambda_tf", sim_res.simulation_parameters.lambda_tf},
+                      {"mu_minus", sim_res.simulation_parameters.mu_minus}}},
                     {"Lowest state energy (eV)", min_energy},
-                    {"Number of stable states", sim_result_100.charge_distributions.size()},
+                    {"Number of stable states", sim_res.charge_distributions.size()},
                     {"Iteration steps",
-                     std::any_cast<uint64_t>(sim_result_100.additional_simulation_parameters.at("iteration_steps"))},
-                    {"alpha", std::any_cast<double>(sim_result_100.additional_simulation_parameters.at("alpha"))}};
+                     std::any_cast<uint64_t>(sim_res.additional_simulation_parameters.at("iteration_steps"))},
+                    {"alpha", std::any_cast<double>(sim_res.additional_simulation_parameters.at("alpha"))}};
             }
+
+            const auto& sim_res = std::get<sim_result_111>(sim_result);
+
             return nlohmann::json{
-                {"Algorithm name", sim_result_111.algorithm_name},
-                {"Simulation runtime", sim_result_111.simulation_runtime.count()},
+                {"Algorithm name", sim_res.algorithm_name},
+                {"Simulation runtime", sim_res.simulation_runtime.count()},
                 {"Physical parameters",
-                 {{"epsilon_r", sim_result_111.simulation_parameters.epsilon_r},
-                  {"lambda_tf", sim_result_111.simulation_parameters.lambda_tf},
-                  {"mu_minus", sim_result_111.simulation_parameters.mu_minus}}},
+                 {{"epsilon_r", sim_res.simulation_parameters.epsilon_r},
+                  {"lambda_tf", sim_res.simulation_parameters.lambda_tf},
+                  {"mu_minus", sim_res.simulation_parameters.mu_minus}}},
                 {"Lowest state energy (eV)", min_energy},
-                {"Number of stable states", sim_result_111.charge_distributions.size()},
+                {"Number of stable states", sim_res.charge_distributions.size()},
                 {"Iteration steps",
-                 std::any_cast<uint64_t>(sim_result_111.additional_simulation_parameters.at("iteration_steps"))},
-                {"alpha", std::any_cast<double>(sim_result_111.additional_simulation_parameters.at("alpha"))}};
+                 std::any_cast<uint64_t>(sim_res.additional_simulation_parameters.at("iteration_steps"))},
+                {"alpha", std::any_cast<double>(sim_res.additional_simulation_parameters.at("alpha"))}};
         }
         catch (...)
         {
+            // If something is off in the variant or additional parameters are missing, return empty JSON.
             return nlohmann::json{};
         }
     }
+
     /**
      * Resets the parameters to their default values.
      */
     void reset_params()
     {
         physical_params = fiction::sidb_simulation_parameters{2, -0.32, 5.6, 5.0};
-        params          = {};
+        qs_params       = {};
+        sim_result      = {};
     }
 };
 
