@@ -6,6 +6,7 @@
 #define FICTION_CMD_TEMP_HPP
 
 #include <fiction/algorithms/simulation/sidb/critical_temperature.hpp>
+#include <fiction/algorithms/simulation/sidb/sidb_simulation_engine.hpp>
 #include <fiction/traits.hpp>
 #include <fiction/types.hpp>
 #include <fiction/utils/math_utils.hpp>
@@ -23,7 +24,7 @@
 namespace alice
 {
 /**
- *
+ * Executes temperature simulation for the current SiDB cell-level layout in store.
  */
 class temp_command : public command
 {
@@ -58,6 +59,12 @@ class temp_command : public command
                    true);
         add_option("--lambda_tf,-l", physical_params.lambda_tf, "Thomas-Fermi screening distance (unit: nm)", true);
         add_option("--mu_minus,-m", physical_params.mu_minus, "Energy transition level (0/-) (unit: eV)", true);
+        add_option("--base", physical_params.base,
+                   "The simulation base, can be 2 or 3, though base-3 critical temperature simulation is experimental "
+                   "(only ClusterComplete supports base-3 simulation)",
+                   true);
+        add_option("--engine", sim_engine_str,
+                   "The simulation engine to use {QuickExact [default], ClusterComplete, QuickSim}", true);
     }
 
   protected:
@@ -72,19 +79,19 @@ class temp_command : public command
 
         if (params.confidence_level <= 0 || params.confidence_level > 1)
         {
-            env->out() << "[e] confidence_level must be in (0, 1]" << std::endl;
+            env->out() << "[e] confidence_level must be in (0, 1]\n";
             reset_params();
             return;
         }
         if (physical_params.epsilon_r <= 0)
         {
-            env->out() << "[e] epsilon_r must be positive" << std::endl;
+            env->out() << "[e] epsilon_r must be positive\n";
             reset_params();
             return;
         }
         if (physical_params.lambda_tf <= 0)
         {
-            env->out() << "[e] lambda_tf must be positive" << std::endl;
+            env->out() << "[e] lambda_tf must be positive\n";
             reset_params();
             return;
         }
@@ -94,7 +101,7 @@ class temp_command : public command
         // error case: empty cell layout store
         if (cs.empty())
         {
-            env->out() << "[w] no cell layout in store" << std::endl;
+            env->out() << "[w] no cell layout in store\n";
             reset_params();
             return;
         }
@@ -106,7 +113,7 @@ class temp_command : public command
         {
             if (ts.empty())
             {
-                env->out() << "[w] no truth table in store" << std::endl;
+                env->out() << "[w] no truth table in store\n";
                 reset_params();
                 return;
             }
@@ -118,65 +125,68 @@ class temp_command : public command
         {
             using Lyt = typename std::decay_t<decltype(lyt_ptr)>::element_type;
 
-            if constexpr (fiction::has_sidb_technology_v<Lyt>)
+            if constexpr (!fiction::has_sidb_technology_v<Lyt>)
             {
-                if constexpr (fiction::is_charge_distribution_surface_v<Lyt>)
+                env->out() << fmt::format("[e] '{}' is not an SiDB layout\n", get_name(lyt_ptr));
+                return;
+            }
+
+            if constexpr (fiction::is_charge_distribution_surface_v<Lyt>)
+            {
+                env->out() << fmt::format(
+                    "[w] '{}' already possesses a charge distribution; no simulation is conducted\n",
+                    get_name(lyt_ptr));
+                return;
+            }
+
+            const auto sim_engine = fiction::get_sidb_simulation_engine(sim_engine_str);
+
+            if (!sim_engine.has_value())
+            {
+                env->out() << fmt::format("[e] '{}' is not a supported SiDB simulation engine\n", sim_engine_str);
+                return;
+            }
+
+            params.operational_params.sim_engine = sim_engine.value();
+
+            params.operational_params.simulation_parameters = physical_params;
+
+            // To aid the compiler
+            if constexpr (fiction::has_sidb_technology_v<Lyt> && !fiction::is_charge_distribution_surface_v<Lyt>)
+            {
+                if (is_set("gate_based"))
                 {
-                    env->out() << fmt::format(
-                                      "[w] {} already possesses a charge distribution; no simulation is conducted",
-                                      get_name(lyt_ptr))
-                               << std::endl;
+                    if (lyt_ptr->num_pis() == 0 || lyt_ptr->num_pos() == 0)
+                    {
+                        env->out() << fmt::format("[e] '{}' requires primary input and output cells to simulate its "
+                                                  "Boolean function\n",
+                                                  get_name(lyt_ptr));
+                        return;
+                    }
+
+                    const auto tt_ptr = ts.current();
+
+                    ct = fiction::critical_temperature_gate_based(*lyt_ptr, std::vector{*tt_ptr}, params, &stats);
                 }
                 else
                 {
-                    params.simulation_parameters = physical_params;
-
-                    if (is_set("gate_based"))
-                    {
-                        if (lyt_ptr->num_pis() == 0 || lyt_ptr->num_pos() == 0)
-                        {
-                            env->out() << fmt::format("[e] {} requires primary input and output cells to simulate its "
-                                                      "Boolean function",
-                                                      get_name(lyt_ptr))
-                                       << std::endl;
-                            return;
-                        }
-
-                        const auto tt_ptr = ts.current();
-
-                        ct = fiction::critical_temperature_gate_based(*lyt_ptr, std::vector<fiction::tt>{*tt_ptr},
-                                                                      params, &stats);
-                    }
-                    else
-                    {
-                        ct = fiction::critical_temperature_non_gate_based(*lyt_ptr, params, &stats);
-                    }
-
-                    if (stats.num_valid_lyt == 0)
-                    {
-                        env->out() << fmt::format("[e] ground state of {} could not be determined", get_name(lyt_ptr))
-                                   << std::endl;
-                    }
-                    else
-                    {
-                        env->out() << fmt::format("[i] critical temperature of {} is {}{} K", get_name(lyt_ptr),
-                                                  (ct == params.max_temperature ? "> " : ""), ct)
-                                   << std::endl;
-
-                        if (stats.num_valid_lyt > 1)
-                        {
-                            env->out() << fmt::format(
-                                              "[i] energy between the ground state and the first erroneous is {} meV",
-                                              fiction::round_to_n_decimal_places(
-                                                  stats.energy_between_ground_state_and_first_erroneous, 2))
-                                       << std::endl;
-                        }
-                    }
+                    ct = fiction::critical_temperature_non_gate_based(*lyt_ptr, params, &stats);
                 }
-            }
-            else
-            {
-                env->out() << fmt::format("[e] {} is not an SiDB layout", get_name(lyt_ptr)) << std::endl;
+
+                if (stats.num_valid_lyt == 0)
+                {
+                    env->out() << fmt::format("[e] ground state of '{}' could not be determined\n", get_name(lyt_ptr));
+                }
+
+                env->out() << fmt::format("[i] critical temperature of '{}' is {}{} K\n", get_name(lyt_ptr),
+                                          (ct == params.max_temperature ? "> " : ""), ct);
+
+                if (stats.num_valid_lyt > 1)
+                {
+                    env->out() << fmt::format(
+                        "[i] energy between the ground state and the first erroneous is {} meV\n",
+                        fiction::round_to_n_decimal_places(stats.energy_between_ground_state_and_first_erroneous, 2));
+                }
             }
         };
 
@@ -202,7 +212,10 @@ class temp_command : public command
      * Critical temperature.
      */
     double ct = 0.0;
-
+    /**
+     * The simulation engine to use.
+     */
+    std::string sim_engine_str{"QuickExact"};
     /**
      * Logs the resulting information in a log file.
      *
