@@ -6,15 +6,14 @@ import sys
 
 import matplotlib.pyplot as plt
 import numpy as np
-
-np.set_printoptions(threshold=sys.maxsize)
 from matplotlib.axes._axes import _log as matplotlib_axes_logger
 
-matplotlib_axes_logger.setLevel('ERROR')
+np.set_printoptions(threshold=sys.maxsize)
+matplotlib_axes_logger.setLevel("ERROR")
 
 
-def rand_int(min, max):
-    return random.randit(min, max)
+def rand_int(minimum: int, maximum: int) -> int:
+    return random.randit(minimum, maximum)
 
 
 class defect_surface:
@@ -23,7 +22,7 @@ class defect_surface:
     # surface_width is the number of dimers within a row
     # surface_height is the number of dimer rows (1 dimer is made of 2 HSi atoms)
     # self.surface_lattice is the array of the surface
-    def __init__(self, surface_width=100, surface_height=100):
+    def __init__(self, surface_width: int = 100, surface_height: int = 100) -> None:
         self.defects_name = None
         self.defect_params = None
         self.defect_plotting = None
@@ -45,59 +44,74 @@ class defect_surface:
     # width and height are how many HSi atoms are used (note this different than surface_height in init)
     # ratio is given as fractional percent (0.05 = 5%)
     # Following
-    def add_defects(self, coverage=0.05):
-        self.defects_name = ['H-Si', 'DB', 'si_vacancy', 'dihydride_pair', 'single_dihydride', 'onebyone', 'threebyone',
-                             'siloxane', 'raised_Si', 'etch_pit', 'missing_dimer']
-        self.defect_params = [[0, 1, 1, 1.],  # HSi
-                              [1, 1, 1, 0.0],  # [array_value,width,height,ratio] DB
-                              [2, 1, 2, 0.0],  # si_vacancy 0.05
-                              [3, 1, 2, 0.2],  # dihydride_pair
-                              [4, 1, 2, 0.05],  # single_dihydride
-                              [5, 4, 2, 0.05],  # onebyone
-                              [6, 4, 4, 0.05],  # threebyone
-                              [7, 1, 2, 0.1],  # siloxane
-                              [8, 1, 2, 0.1],  # raised_si
-                              [9, 3, 2, 0.05],  # etch pit
-                              [10, 1, 2, 0.05]  # missing_dimer
-                              ]
+    def add_defects(self, coverage: float = 0.05) -> None:
+        self.defects_name = [
+            "H-Si",
+            "DB",
+            "si_vacancy",
+            "dihydride_pair",
+            "single_dihydride",
+            "onebyone",
+            "threebyone",
+            "siloxane",
+            "raised_Si",
+            "etch_pit",
+            "missing_dimer",
+        ]
+        self.defect_params = [
+            [0, 1, 1, 1.0],  # HSi
+            [1, 1, 1, 0.0],  # [array_value,width,height,ratio] DB
+            [2, 1, 2, 0.0],  # si_vacancy 0.05
+            [3, 1, 2, 0.2],  # dihydride_pair
+            [4, 1, 2, 0.05],  # single_dihydride
+            [5, 4, 2, 0.05],  # onebyone
+            [6, 4, 4, 0.05],  # threebyone
+            [7, 1, 2, 0.1],  # siloxane
+            [8, 1, 2, 0.1],  # raised_si
+            [9, 3, 2, 0.05],  # etch pit
+            [10, 1, 2, 0.05],  # missing_dimer
+        ]
 
-        self.defect_plotting = [[(255 / 255, 0, 0), 'o'],  # HSi
-                                [(0, 0, 255 / 255), 'o'],  # [colour,shape] DB
-                                [(0, 0, 0), 's'],  # si_vacancy
-                                [(58 / 255, 0, 83 / 255), 's'],  # dihydride_pair
-                                [(72 / 255, 53 / 255, 4 / 255), 's'],  # single_dihydride
-                                [(42 / 255, 56 / 255, 14 / 255), 's'],  # onebyone
-                                [(173 / 255, 255 / 255, 47 / 255), 's'],  # threebyone
-                                [(64 / 255, 224 / 255, 208 / 255), 's'],  # siloxane
-                                [(119 / 255, 136 / 255, 153 / 255), 's'],  # raised_si
-                                [(255 / 255, 255 / 255, 0), 's'],  # etch_pit
-                                [(227 / 255, 28 / 255, 121 / 255), 's']  # missing_dimer
-                                ]
+        self.defect_plotting = [
+            [(255 / 255, 0, 0), "o"],  # HSi
+            [(0, 0, 255 / 255), "o"],  # [colour,shape] DB
+            [(0, 0, 0), "s"],  # si_vacancy
+            [(58 / 255, 0, 83 / 255), "s"],  # dihydride_pair
+            [(72 / 255, 53 / 255, 4 / 255), "s"],  # single_dihydride
+            [(42 / 255, 56 / 255, 14 / 255), "s"],  # onebyone
+            [(173 / 255, 255 / 255, 47 / 255), "s"],  # threebyone
+            [(64 / 255, 224 / 255, 208 / 255), "s"],  # siloxane
+            [(119 / 255, 136 / 255, 153 / 255), "s"],  # raised_si
+            [(255 / 255, 255 / 255, 0), "s"],  # etch_pit
+            [(227 / 255, 28 / 255, 121 / 255), "s"],  # missing_dimer
+        ]
 
         self.total_defect_lattice_points = int(self.surface_width * self.surface_height * coverage)
         for defect in self.defect_params:
             num_of_defects = int(defect[3] * self.total_defect_lattice_points)
             for i in range(1, num_of_defects):
-
                 random_width = random.randint(0, self.surface_width - 1)
                 random_height = random.randint(0, self.surface_height - 1)
 
                 if (random_width > random_width + defect[1] and random_height > random_height + defect[2]) or (
-                        random_height % 2 == 1 and defect[2] % 2 == 0):
+                    random_height % 2 == 1 and defect[2] % 2 == 0
+                ):
                     i = i - 1
-                    pass
                 else:
-
-                    if np.all(self.surface_lattice[random_height:random_height + defect[2],
-                              random_width:random_width + defect[1]] == 0):
-                        self.surface_lattice[random_height:random_height + defect[2],
-                        random_width:random_width + defect[1]] = defect[0]
+                    if np.all(
+                        self.surface_lattice[
+                            random_height : random_height + defect[2], random_width : random_width + defect[1]
+                        ]
+                        == 0
+                    ):
+                        self.surface_lattice[
+                            random_height : random_height + defect[2], random_width : random_width + defect[1]
+                        ] = defect[0]
 
                     else:
                         i = i - 1
-                        pass
 
-    def draw_panels(self):  # DB_panels,DB_pattern_extended, pattern):
+    def draw_panels(self) -> None:  # DB_panels,DB_pattern_extended, pattern):
         # draws the DB_pattern_extended with rectangles to show each pannel
 
         width_nm = self.a1 * self.surface_width
@@ -108,40 +122,45 @@ class defect_surface:
         # DB_bottom_points = np.where(self.surface_lattice==1)
         # print(lattice_points)
         fig = plt.figure(figsize=((width_nm + 1) / 10, (height_nm + 1) / 10), dpi=100)
-        ax = fig.add_subplot(1, 1, 1)
+        fig.add_subplot(1, 1, 1)
         plt.gca().invert_yaxis()
 
         label_list = np.full((np.shape(self.defect_params)[0]), False)
 
-        for y in range(0, self.surface_height):
-            for x in range(0, self.surface_width):
+        for y in range(self.surface_height):
+            for x in range(self.surface_width):
                 if not label_list[self.surface_lattice[y][x]]:
                     lab = self.defects_name[self.surface_lattice[y][x]]
                     label_list[self.surface_lattice[y][x]] = True
                 else:
-                    lab = ''
+                    lab = ""
 
                 if y % 2 == 0:
-                    plt.scatter(x * self.a1, y * self.a2 + self.b2 / 2,
-                                c=self.defect_plotting[self.surface_lattice[y][x]][0],
-                                label=lab,
-                                marker=self.defect_plotting[self.surface_lattice[y][x]][1])
+                    plt.scatter(
+                        x * self.a1,
+                        y * self.a2 + self.b2 / 2,
+                        c=self.defect_plotting[self.surface_lattice[y][x]][0],
+                        label=lab,
+                        marker=self.defect_plotting[self.surface_lattice[y][x]][1],
+                    )
                 elif y % 2 == 1:
-                    plt.scatter(x * self.a1, y * self.a2 - self.b2 / 2,
-                                c=self.defect_plotting[self.surface_lattice[y][x]][0],
-                                label=lab,
-                                marker=self.defect_plotting[self.surface_lattice[y][x]][1])
+                    plt.scatter(
+                        x * self.a1,
+                        y * self.a2 - self.b2 / 2,
+                        c=self.defect_plotting[self.surface_lattice[y][x]][0],
+                        label=lab,
+                        marker=self.defect_plotting[self.surface_lattice[y][x]][1],
+                    )
 
-        plt.axis('equal')
+        plt.axis("equal")
         plt.xticks([])
         plt.yticks([])
         plt.legend()
         # print(self.DB_pattern_extended.shape)
         plt.show()
 
-    def save_to_file(self, filename='test.txt'):
+    def save_to_file(self, filename: str = "test.txt") -> None:
         np.savetxt(filename, self.surface_lattice)
-        print('file_saved')
 
 
 surface_width = 740
@@ -149,6 +168,5 @@ surface_height = 1090
 coverage = 0.005
 surface = defect_surface(surface_width=surface_width, surface_height=surface_height)
 surface.add_defects(coverage=coverage)
-print(surface.surface_lattice)
 # surface.draw_panels()
 # surface.save_to_file('test.txt')
