@@ -39,7 +39,7 @@ int main()  // NOLINT
         "QuickSim Single Runtime [s]",
         "Total QuickSim TTS [s]"};
 
-    static const std::string folder = fmt::format("{}bestagon_gates_type_tags/", EXPERIMENTS_PATH);
+    static const std::string folder = fmt::format("{}sidb_gate_libraries/bestagon_gates/", EXPERIMENTS_PATH);
 
     static const std::array<std::pair<std::string, std::vector<tt>>, 12> gates = {
         std::make_pair("hourglass", create_crossing_wire_tt()),
@@ -62,55 +62,52 @@ int main()  // NOLINT
 
     for (const auto& [gate, truth_table] : gates)
     {
-        for (const auto& file : std::filesystem::directory_iterator(fmt::format("{}{}", folder, gate)))
+        const auto layout = read_sqd_layout<sidb_100_cell_clk_lyt_siqad>(fmt::format("{}{}.sqd", folder, gate));
+
+        double      runtime_exhaustive      = 0;
+        double      runtime_quickexact      = 0;
+        double      tts_quicksim            = 0;
+        std::size_t instances               = 0;
+        double      quicksim_accuracy_mean  = 0.0;
+        double      quicksim_single_runtime = 0.0;
+
+        // simulate layout with no input pattern
+        const auto             exhaustive_results_layout = exhaustive_ground_state_simulation(layout, sim_params);
+        time_to_solution_stats stats{};
+        time_to_solution(layout, qs_params, tts_params, &stats);
+        const auto quickexact_results_layout = quickexact(layout, qe_params);
+
+        runtime_exhaustive += mockturtle::to_seconds(exhaustive_results_layout.simulation_runtime);
+        runtime_quickexact += mockturtle::to_seconds(quickexact_results_layout.simulation_runtime);
+        tts_quicksim += stats.time_to_solution;
+        instances += 1;
+        quicksim_accuracy_mean += stats.acc;
+        quicksim_single_runtime += stats.mean_single_runtime;
+
+        // simulate layout with all input patterns
+        auto       bii                = bdl_input_iterator<sidb_100_cell_clk_lyt_siqad>{layout};
+        const auto num_input_patterns = truth_table.front().num_bits();
+
+        for (auto i = 0u; i < num_input_patterns; ++i, ++bii)
         {
-            const auto layout = read_sqd_layout<sidb_100_cell_clk_lyt_siqad>(file.path().string());
+            const auto             exhaustive_results = exhaustive_ground_state_simulation(*bii, sim_params);
+            time_to_solution_stats tts_stats{};
+            time_to_solution(*bii, qs_params, tts_params, &tts_stats);
+            const auto quickexact_results = quickexact(*bii, qe_params);
 
-            double      runtime_exhaustive      = 0;
-            double      runtime_quickexact      = 0;
-            double      tts_quicksim            = 0;
-            std::size_t instances               = 0;
-            double      quicksim_accuracy_mean  = 0.0;
-            double      quicksim_single_runtime = 0.0;
-
-            // simulate layout with no input pattern
-            const auto             exhaustive_results_layout = exhaustive_ground_state_simulation(layout, sim_params);
-            time_to_solution_stats stats{};
-            time_to_solution(layout, qs_params, tts_params, &stats);
-            const auto quickexact_results_layout = quickexact(layout, qe_params);
-
-            runtime_exhaustive += mockturtle::to_seconds(exhaustive_results_layout.simulation_runtime);
-            runtime_quickexact += mockturtle::to_seconds(quickexact_results_layout.simulation_runtime);
+            runtime_exhaustive += mockturtle::to_seconds(exhaustive_results.simulation_runtime);
+            runtime_quickexact += mockturtle::to_seconds(quickexact_results.simulation_runtime);
             tts_quicksim += stats.time_to_solution;
+
             instances += 1;
             quicksim_accuracy_mean += stats.acc;
             quicksim_single_runtime += stats.mean_single_runtime;
-
-            // simulate layout with all input patterns
-            auto       bii                = bdl_input_iterator<sidb_100_cell_clk_lyt_siqad>{layout};
-            const auto num_input_patterns = truth_table.front().num_bits();
-
-            for (auto i = 0u; i < num_input_patterns; ++i, ++bii)
-            {
-                const auto             exhaustive_results = exhaustive_ground_state_simulation(*bii, sim_params);
-                time_to_solution_stats tts_stats{};
-                time_to_solution(*bii, qs_params, tts_params, &tts_stats);
-                const auto quickexact_results = quickexact(*bii, qe_params);
-
-                runtime_exhaustive += mockturtle::to_seconds(exhaustive_results.simulation_runtime);
-                runtime_quickexact += mockturtle::to_seconds(quickexact_results.simulation_runtime);
-                tts_quicksim += stats.time_to_solution;
-
-                instances += 1;
-                quicksim_accuracy_mean += stats.acc;
-                quicksim_single_runtime += stats.mean_single_runtime;
-            }
-            quicksim_accuracy_mean = quicksim_accuracy_mean / static_cast<double>(instances);
-            simulation_exp(gate, instances, runtime_exhaustive, runtime_quickexact, quicksim_accuracy_mean,
-                           quicksim_single_runtime, tts_quicksim);
-            simulation_exp.save();
-            simulation_exp.table();
         }
+        quicksim_accuracy_mean = quicksim_accuracy_mean / static_cast<double>(instances);
+        simulation_exp(gate, instances, runtime_exhaustive, runtime_quickexact, quicksim_accuracy_mean,
+                       quicksim_single_runtime, tts_quicksim);
+        simulation_exp.save();
+        simulation_exp.table();
     }
     return EXIT_SUCCESS;
 }
