@@ -30,7 +30,7 @@ int main()  // NOLINT
         "E_{g,err} [meV]",
     };
 
-    static const std::string folder = fmt::format("{}bestagon_gates_type_tags/", EXPERIMENTS_PATH);
+    static const std::string folder = fmt::format("{}sidb_gate_libraries/bestagon_gates/", EXPERIMENTS_PATH);
 
     static const std::array<std::pair<std::string, std::vector<tt>>, 12> gates = {
         std::make_pair("hourglass", create_double_wire_tt()),
@@ -51,19 +51,16 @@ int main()  // NOLINT
 
     for (const auto& [gate, truth_table] : gates)
     {
-        for (const auto& file : std::filesystem::directory_iterator(fmt::format("{}{}", folder, gate)))
-        {
-            const auto layout = read_sqd_layout<sidb_100_cell_clk_lyt_siqad>(file.path().string());
+        const auto layout = read_sqd_layout<sidb_100_cell_clk_lyt_siqad>(fmt::format("{}/{}.sqd", folder, gate));
 
-            critical_temperature_stats ct_stats{};
+        critical_temperature_stats ct_stats{};
 
-            const auto ct =
-                critical_temperature_gate_based<sidb_100_cell_clk_lyt_siqad>(layout, truth_table, ct_params, &ct_stats);
+        const auto ct =
+            critical_temperature_gate_based<sidb_100_cell_clk_lyt_siqad>(layout, truth_table, ct_params, &ct_stats);
 
-            simulation_exp(gate, ct, ct_stats.energy_between_ground_state_and_first_erroneous);
-            simulation_exp.save();
-            simulation_exp.table();
-        }
+        simulation_exp(gate, ct, ct_stats.energy_between_ground_state_and_first_erroneous);
+        simulation_exp.save();
+        simulation_exp.table();
     }
     return EXIT_SUCCESS;
 }
