@@ -10,6 +10,7 @@
 #include "fiction/technology/cell_technologies.hpp"
 #include "fiction/traits.hpp"
 
+#include <cassert>
 #include <cstdint>
 #include <iterator>
 #include <vector>
@@ -76,17 +77,17 @@ class bdl_input_iterator
      * BDL pairs.
      *
      * @param lyt The SiDB BDL layout to iterate over.
-     * @param params Parameters for the BDL input iterator.
+     * @param ps Parameters for the BDL input iterator.
      */
     explicit bdl_input_iterator(const Lyt&                       lyt,
-                                const bdl_input_iterator_params& params = bdl_input_iterator_params{}) noexcept :
+                                const bdl_input_iterator_params& ps = bdl_input_iterator_params{}) noexcept :
             layout{lyt.clone()},
             input_pairs{
-                detect_bdl_pairs<Lyt>(lyt, sidb_technology::cell_type::INPUT, params.bdl_wire_params.bdl_pairs_params)},
+                detect_bdl_pairs<Lyt>(lyt, sidb_technology::cell_type::INPUT, ps.bdl_wire_params.bdl_pairs_params)},
             num_inputs{static_cast<uint8_t>(input_pairs.size())},
-            input_bdl_wires{detect_bdl_wires<Lyt>(lyt, params.bdl_wire_params, bdl_wire_selection::INPUT)},
+            input_bdl_wires{detect_bdl_wires<Lyt>(lyt, ps.bdl_wire_params, bdl_wire_selection::INPUT)},
             last_bdl_for_each_wire{determine_last_bdl_for_each_wire()},
-            params{params}
+            params{ps}
     {
         static_assert(is_cell_level_layout_v<Lyt>, "Lyt is not a cell-level layout");
         static_assert(has_sidb_technology_v<Lyt>, "Lyt is not an SiDB layout");
@@ -98,18 +99,18 @@ class bdl_input_iterator
      * which assigns binary `0` to all input BDL pairs.
      *
      * @param lyt The SiDB BDL layout to iterate over.
-     * @param params Parameters for the BDL input iterator.
+     * @param ps Parameters for the BDL input iterator.
      * @param input_wires Pre-detected input BDL wires.
      */
-    explicit bdl_input_iterator(const Lyt& lyt, const bdl_input_iterator_params& params,
+    explicit bdl_input_iterator(const Lyt& lyt, const bdl_input_iterator_params& ps,
                                 const std::vector<bdl_wire<Lyt>>& input_wires) noexcept :
             layout{lyt.clone()},
             input_pairs{
-                detect_bdl_pairs<Lyt>(lyt, sidb_technology::cell_type::INPUT, params.bdl_wire_params.bdl_pairs_params)},
+                detect_bdl_pairs<Lyt>(lyt, sidb_technology::cell_type::INPUT, ps.bdl_wire_params.bdl_pairs_params)},
             num_inputs{static_cast<uint8_t>(input_pairs.size())},
             input_bdl_wires{input_wires},
             last_bdl_for_each_wire{determine_last_bdl_for_each_wire()},
-            params{params}
+            params{ps}
     {
         static_assert(is_cell_level_layout_v<Lyt>, "Lyt is not a cell-level layout");
         static_assert(has_sidb_technology_v<Lyt>, "Lyt is not an SiDB layout");
@@ -340,6 +341,15 @@ class bdl_input_iterator
     {
         return input_pairs.size();
     }
+    /**
+     * Returns the current input index.
+     *
+     * @return The current input index.
+     */
+    [[nodiscard]] uint64_t get_current_input_index() const noexcept
+    {
+        return current_input_index;
+    }
 
   private:
     /**
@@ -425,6 +435,8 @@ class bdl_input_iterator
      */
     void set_all_inputs() noexcept
     {
+        assert(num_inputs == input_bdl_wires.size() && "number of inputs and number of wires don't match");
+
         for (uint64_t i = num_inputs - 1; i < num_inputs; --i)
         {
             const auto& input_i = input_pairs[i];
