@@ -10,10 +10,8 @@
 #include "fiction/algorithms/simulation/sidb/is_operational.hpp"
 #include "fiction/algorithms/simulation/sidb/random_sidb_layout_generator.hpp"
 #include "fiction/algorithms/simulation/sidb/sidb_simulation_engine.hpp"
-#include "fiction/technology/cell_ports.hpp"
 #include "fiction/technology/cell_technologies.hpp"
 #include "fiction/technology/charge_distribution_surface.hpp"
-#include "fiction/technology/physical_constants.hpp"
 #include "fiction/technology/sidb_charge_state.hpp"
 #include "fiction/technology/sidb_defects.hpp"
 #include "fiction/traits.hpp"
@@ -21,18 +19,16 @@
 #include "fiction/utils/math_utils.hpp"
 
 #include <fmt/format.h>
-#include <kitty/dynamic_truth_table.hpp>
 #include <kitty/traits.hpp>
 #include <mockturtle/utils/stopwatch.hpp>
 
 #include <algorithm>
 #include <atomic>
-#include <bitset>
 #include <cassert>
-#include <cmath>
 #include <cstdint>
 #include <cstdlib>
 #include <iostream>
+#include <iterator>
 #include <limits>
 #include <mutex>
 #include <optional>
@@ -140,9 +136,30 @@ struct design_sidb_gates_stats
      */
     mockturtle::stopwatch<>::duration time_total{0};
     /**
+     * The runtime of the pruning process.
+     */
+    mockturtle::stopwatch<>::duration pruning_total{0};
+    /**
      * The simulation engine to be used for the operational domain computation.
      */
     sidb_simulation_engine sim_engine{sidb_simulation_engine::QUICKEXACT};
+    /**
+     * The number of all possible layouts.
+     */
+    std::size_t number_of_layouts{0};
+    /**
+     * The number of layouts that remain after first pruning (discarding layouts with potential positive SiDBs).
+     */
+    std::size_t number_of_layouts_after_first_pruning{0};
+    /**
+     * The number of layouts that remain after second pruning (discarding layouts that fail to satisfy the physical
+     * model).
+     */
+    std::size_t number_of_layouts_after_second_pruning{0};
+    /**
+     * The number of layouts that remain after third pruning (discarding layouts with unstable I/O signals).
+     */
+    std::size_t number_of_layouts_after_third_pruning{0};
     /**
      * This function outputs the total time taken for the SiDB gate design process to the provided output stream.
      * If no output stream is provided, it defaults to standard output (`std::cout`).
@@ -188,7 +205,8 @@ class design_sidb_gates_impl
             number_of_output_wires{output_bdl_wires.size()},
             all_canvas_layouts{determine_all_possible_canvas_layouts()}
     {
-        stats.sim_engine = params.operational_params.sim_engine;
+        stats.number_of_layouts = all_canvas_layouts.size();
+        stats.sim_engine        = params.operational_params.sim_engine;
     }
 
     /**
@@ -248,9 +266,8 @@ class design_sidb_gates_impl
             // canvas SiDBs are added to the skeleton
             const auto layout_with_added_cells = skeleton_layout_with_canvas_sidbs(combination);
 
-            if (const auto [status, aux_stats] =
-                    is_operational(layout_with_added_cells, truth_table, params.operational_params,
-                                   std::optional{input_bdl_wires}, std::optional{output_bdl_wires});
+            if (const auto [status, aux_stats] = is_operational(
+                    layout_with_added_cells, truth_table, params.operational_params, input_bdl_wires, output_bdl_wires);
                 status == operational_status::OPERATIONAL)
             {
                 {
@@ -269,16 +286,14 @@ class design_sidb_gates_impl
             }
         };
 
-        const std::size_t number_of_used_threads =
-            std::min(static_cast<std::size_t>(num_threads), all_combinations.size());
+        const std::size_t num_threads = std::min(number_of_threads, all_combinations.size());
 
-        const std::size_t chunk_size =
-            (all_combinations.size() + number_of_used_threads - 1) / number_of_used_threads;  // Ceiling division
+        const std::size_t chunk_size = (all_combinations.size() + num_threads - 1) / num_threads;  // Ceiling division
 
         std::vector<std::thread> threads{};
-        threads.reserve(number_of_used_threads);
+        threads.reserve(num_threads);
 
-        for (std::size_t i = 0; i < number_of_used_threads; ++i)
+        for (std::size_t i = 0; i < num_threads; ++i)
         {
             threads.emplace_back(
                 [this, i, chunk_size, &all_combinations, &add_combination_to_layout_and_check_operation,
@@ -295,6 +310,7 @@ class design_sidb_gates_impl
                         {
                             return;
                         }
+
                         add_combination_to_layout_and_check_operation(all_combinations[j]);
                     }
                 });
@@ -336,6 +352,8 @@ class design_sidb_gates_impl
             params.canvas, params.number_of_sidbs,
             generate_random_sidb_layout_params<cell<Lyt>>::positive_charges::ALLOWED};
 
+        const auto num_threads = std::min(number_of_threads, all_canvas_layouts.size());
+
         std::vector<std::thread> threads{};
         threads.reserve(num_threads);
 
@@ -352,6 +370,7 @@ class design_sidb_gates_impl
                     while (!gate_layout_is_found)
                     {
                         auto result_lyt = generate_random_sidb_layout<Lyt>(skeleton_layout, parameter);
+
                         if constexpr (has_get_sidb_defect_v<Lyt>)
                         {
                             result_lyt.foreach_sidb_defect(
@@ -363,9 +382,9 @@ class design_sidb_gates_impl
                                     }
                                 });
                         }
-                        if (const auto [status, aux_stats] =
-                                is_operational(result_lyt, truth_table, params.operational_params,
-                                               std::optional{input_bdl_wires}, std::optional{output_bdl_wires});
+
+                        if (const auto [status, aux_stats] = is_operational(
+                                result_lyt, truth_table, params.operational_params, input_bdl_wires, output_bdl_wires);
                             status == operational_status::OPERATIONAL)
                         {
                             const std::lock_guard lock{mutex_to_protect_designed_gate_layouts};
@@ -405,10 +424,24 @@ class design_sidb_gates_impl
      *
      * @return A vector of designed SiDB gate layouts.
      */
-    [[nodiscard]] std::vector<Lyt> run_quickcell() const noexcept
+    [[nodiscard]] std::vector<Lyt> run_quickcell() noexcept
     {
         mockturtle::stopwatch stop{stats.time_total};
-        const auto            gate_candidates = run_pruning();
+
+        std::vector<Lyt> gate_candidates{};
+        gate_candidates.reserve(all_canvas_layouts.size());
+
+        {
+            mockturtle::stopwatch stop_pruning{stats.pruning_total};
+            gate_candidates = run_pruning();
+        }
+
+        stats.number_of_layouts_after_first_pruning =
+            all_canvas_layouts.size() - number_of_discarded_layouts_at_first_pruning.load();
+        stats.number_of_layouts_after_second_pruning =
+            stats.number_of_layouts_after_first_pruning - number_of_discarded_layouts_at_second_pruning.load();
+        stats.number_of_layouts_after_third_pruning =
+            stats.number_of_layouts_after_second_pruning - number_of_discarded_layouts_at_third_pruning.load();
 
         std::vector<Lyt> gate_layouts{};
         gate_layouts.reserve(gate_candidates.size());
@@ -418,21 +451,32 @@ class design_sidb_gates_impl
             return gate_layouts;
         }
 
+        std::optional<std::vector<std::vector<std::vector<charge_distribution_surface<Lyt>>>>>
+            sim_results_per_input_for_each_gate_design{};
+
+        if (params.post_design_process ==
+            design_sidb_gates_params<cell<Lyt>>::post_design_mode::PREFER_ENERGETICALLY_ISOLATED_GROUND_STATES)
+        {
+            sim_results_per_input_for_each_gate_design =
+                std::make_optional<std::vector<std::vector<std::vector<charge_distribution_surface<Lyt>>>>>();
+        }
+
         std::mutex mutex_to_protect_gate_designs{};
 
         gate_layouts.reserve(gate_candidates.size());
 
-        const std::size_t number_of_threads =
-            std::min(static_cast<std::size_t>(std::thread::hardware_concurrency()), gate_candidates.size());
+        const std::size_t num_threads = std::min(number_of_threads, gate_candidates.size());
+
         const std::size_t chunk_size = (gate_candidates.size() + num_threads - 1) / num_threads;  // Ceiling division
 
         std::vector<std::thread> threads;
-        threads.reserve(number_of_threads);
+        threads.reserve(num_threads);
 
         std::atomic<bool> gate_design_found = false;
 
         const auto check_operational_status =
-            [this, &gate_layouts, &mutex_to_protect_gate_designs, &gate_design_found](const auto& candidate) noexcept
+            [this, &gate_layouts, &mutex_to_protect_gate_designs, &gate_design_found,
+             &sim_results_per_input_for_each_gate_design](const auto& candidate) noexcept
         {
             // Early exit if a solution is found and only the first solution is required
             if (gate_design_found && (params.termination_cond ==
@@ -440,21 +484,32 @@ class design_sidb_gates_impl
             {
                 return;
             }
-            if (const auto [status, aux_stats] =
-                    is_operational(candidate, truth_table, params.operational_params, std::optional{input_bdl_wires},
-                                   std::optional{output_bdl_wires});
+
+            // pruning was already conducted above. Hence, SIMULATION_ONLY is chosen.
+            params.operational_params.strategy_to_analyze_operational_status =
+                is_operational_params::operational_analysis_strategy::SIMULATION_ONLY;
+
+            if (const auto [status, aux_stats] = is_operational(candidate, truth_table, params.operational_params,
+                                                                input_bdl_wires, output_bdl_wires);
                 status == operational_status::OPERATIONAL)
             {
                 // Lock and update shared resources
                 {
                     const std::lock_guard lock{mutex_to_protect_gate_designs};
-                    gate_layouts.push_back(candidate);
+
+                    gate_layouts.emplace_back(candidate);
+
+                    if (sim_results_per_input_for_each_gate_design.has_value())
+                    {
+                        sim_results_per_input_for_each_gate_design.value().emplace_back(
+                            std::move(aux_stats.simulation_results.value()));
+                    }
                 }
                 gate_design_found = true;  // Notify all threads that a solution has been found
             }
         };
 
-        for (std::size_t i = 0; i < number_of_threads; ++i)
+        for (std::size_t i = 0; i < num_threads; ++i)
         {
             threads.emplace_back(
                 [this, i, chunk_size, &gate_candidates, &check_operational_status, &gate_design_found]()
@@ -484,6 +539,14 @@ class design_sidb_gates_impl
             }
         }
 
+        if (params.post_design_process ==
+                design_sidb_gates_params<cell<Lyt>>::post_design_mode::PREFER_ENERGETICALLY_ISOLATED_GROUND_STATES &&
+            gate_layouts.size() > 1)
+        {
+            sort_designed_gate_layouts_by_ground_state_isolation(
+                gate_layouts, std::move(sim_results_per_input_for_each_gate_design.value()));
+        }
+
         return gate_layouts;
     }
 
@@ -500,15 +563,11 @@ class design_sidb_gates_impl
     /**
      * Parameters for the *SiDB Gate Designer*.
      */
-    const design_sidb_gates_params<cell<Lyt>> params;
+    design_sidb_gates_params<cell<Lyt>> params;
     /**
      * All cells within the canvas.
      */
     std::vector<typename Lyt::cell> all_sidbs_in_canvas;
-    /**
-     * Number of threads to be used for parallel execution.
-     */
-    const std::size_t num_threads{std::thread::hardware_concurrency()};
     /**
      * The statistics of the gate design.
      */
@@ -534,406 +593,21 @@ class design_sidb_gates_impl
      */
     const std::vector<Lyt> all_canvas_layouts{};
     /**
-     * This function assigns the charge states of the input wires in the layout according to the provided input pattern
-     * index. It performs the following steps:
-     * - For `NORTH-SOUTH` port wires, if the corresponding bit in the input pattern is set, assigns `NEUTRAL`
-     * charge to the upper part and `NEGATIVE` charge to the lower part of the BDLs of the wire.
-     * - For `NORTH-SOUTH` port wires, if the corresponding bit in the input pattern is not set, assigns `NEGATIVE`
-     *   charge to the upper part and `NEUTRAL` charge to the lower part of the BDLs of the wire.
-     * - For `SOUTH-NORTH` port wires, if the corresponding bit in the input pattern is set, assigns `NEGATIVE`
-     * charge to the upper part and `NEUTRAL` charge to the lower part of the BDLs of the wire.
-     * - For `SOUTH-NORTH` port wires, if the corresponding bit in the input pattern is not set, assigns `NEUTRAL`
-     *   charge to the upper part and `NEGATIVE` charge to the lower part of the BDLs of the wire.
-     *
-     * @param layout The charge distribution surface layout to be modified.
-     * @param current_input_index The index representing the current input pattern.
+     * Number of discarded layouts at first pruning.
      */
-    void
-    set_charge_distribution_of_input_wires_based_on_input_pattern(charge_distribution_surface<Lyt>& layout,
-                                                                  const uint64_t current_input_index) const noexcept
-    {
-        layout.assign_all_charge_states(sidb_charge_state::NEGATIVE, charge_index_mode::KEEP_CHARGE_INDEX);
-
-        for (auto i = 0u; i < number_of_input_wires; i++)
-        {
-            if (input_bdl_wires[number_of_input_wires - 1 - i].port.dir == port_direction::SOUTH)
-            {
-                if ((current_input_index & (uint64_t{1ull} << i)) != 0ull)
-                {
-                    for (const auto& bdl : input_bdl_wires[number_of_input_wires - 1 - i].pairs)
-                    {
-                        if (bdl.type == sidb_technology::INPUT)
-                        {
-                            continue;
-                        }
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-                else
-                {
-                    for (const auto& bdl : input_bdl_wires[number_of_input_wires - 1 - i].pairs)
-                    {
-                        if (bdl.type == sidb_technology::INPUT)
-                        {
-                            continue;
-                        }
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-            }
-            else if (input_bdl_wires[number_of_input_wires - 1 - i].port.dir == port_direction::NORTH)
-            {
-                if ((current_input_index & (uint64_t{1ull} << i)) != 0ull)
-                {
-                    for (const auto& bdl : input_bdl_wires[number_of_input_wires - 1 - i].pairs)
-                    {
-                        if (bdl.type == sidb_technology::INPUT)
-                        {
-                            continue;
-                        }
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-                else
-                {
-                    for (const auto& bdl : input_bdl_wires[number_of_input_wires - 1 - i].pairs)
-                    {
-                        if (bdl.type == sidb_technology::INPUT)
-                        {
-                            continue;
-                        }
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-            }
-        }
-    }
-
+    std::atomic<std::size_t> number_of_discarded_layouts_at_first_pruning{0};
     /**
-     * This function assigns the charge states of the input wires in the layout according to the provided input pattern
-     * index. It performs the following steps:
-     *    - For `NORTH-SOUTH` port wires, if the corresponding bit in the input pattern is set, assigns `NEUTRAL`
-     * charge to the upper part and `NEGATIVE` charge to the lower part of the BDLs of the wire.
-     *    - For `NORTH-SOUTH` port wires, if the corresponding bit in the input pattern is not set, assigns
-     * `NEGATIVE` charge to the upper part and `NEUTRAL` charge to the lower part of the BDLs of the wire.
-     *    - For `SOUTH-NORTH` port wires, if the corresponding bit in the input pattern is set, assigns `NEGATIVE`
-     * charge to the upper part and `NEUTRAL` charge to the lower part of the BDLs of the wire.
-     *    - For `SOUTH-NORTH` port wires, if the corresponding bit in the input pattern is not set, assigns
-     * `NEUTRAL` charge to the upper part and `NEGATIVE` charge to the lower part of the BDLs of the wire.
-     *
-     * @param layout The charge distribution surface layout to be modified.
-     * @param current_input_index The index representing the current input pattern.
-     * @return void
+     * Number of discarded layouts at second pruning.
      */
-    void set_charge_distribution_of_output_wires_based_on_output_index(charge_distribution_surface<Lyt>& layout,
-                                                                       const uint64_t output_wire_index) const noexcept
-    {
-        for (auto i = 0u; i < number_of_output_wires; i++)
-        {
-            if (output_bdl_wires[i].port.dir == port_direction::SOUTH)
-            {
-                if ((output_wire_index & (uint64_t{1ull} << i)) != 0ull)
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-                else
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-            }
-            else if (output_bdl_wires[i].port.dir == port_direction::NORTH)
-            {
-                if ((output_wire_index & (uint64_t{1ull} << i)) != 0ull)
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        if (bdl.type == sidb_technology::INPUT)
-                        {
-                            continue;
-                        }
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-                else
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-            }
-        }
-    }
-
+    std::atomic<std::size_t> number_of_discarded_layouts_at_second_pruning{0};
     /**
-     * This function calculates the output index for a given input index by evaluating the truth table.
-     *
-     * @param current_input_index The index representing the current input pattern.
-     * @return The output index derived from the truth table for the given input index.
+     * Number of discarded layouts at third pruning.
      */
-    [[nodiscard]] uint64_t determine_output_index_of_output(const uint64_t current_input_index) const noexcept
-    {
-        std::bitset<64> bits{};
-        for (auto i = 0u; i < number_of_output_wires; i++)
-        {
-            bits[i] = kitty::get_bit(truth_table[i], current_input_index);
-        }
-        return bits.to_ulong();
-    }
-
+    std::atomic<std::size_t> number_of_discarded_layouts_at_third_pruning{0};
     /**
-     * This function assigns the charge states of the output wires in the layout according to the values in the truth
-     * table for the provided input pattern index.
-     *
-     * @param layout The charge distribution surface layout to be modified.
-     * @param input_index The index representing the current input pattern.
+     * Number of threads to be used for the design process.
      */
-    void set_charge_distribution_of_output_wires_based_on_truth_table(charge_distribution_surface<Lyt>& layout,
-                                                                      const uint64_t input_index) const noexcept
-    {
-        for (auto i = 0u; i < number_of_output_wires; i++)
-        {
-            if (output_bdl_wires[i].port.dir == port_direction::SOUTH)
-            {
-                if (kitty::get_bit(truth_table[i], input_index))
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-                else
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-            }
-            else if (output_bdl_wires[i].port.dir == port_direction::NORTH)
-            {
-                if (kitty::get_bit(truth_table[i], input_index))
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-                else
-                {
-                    for (const auto& bdl : output_bdl_wires[i].pairs)
-                    {
-                        layout.assign_charge_state(bdl.upper, sidb_charge_state::NEUTRAL,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                        layout.assign_charge_state(bdl.lower, sidb_charge_state::NEGATIVE,
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * This function determines if there is a charge distribution of the canvas SiDBs for which the charge distribution
-     * of the whole layout is physically valid.
-     *
-     * @param cds_layout The charge distribution surface layout to be evaluated.
-     * @param cds_canvas The charge distribution surface of the canvas SiDBs. All possible configurations are enumerated
-     * @return The minimum energy value if a physically valid configuration is found, `std::nullopt`
-     * otherwise.
-     */
-    [[nodiscard]] std::optional<double>
-    is_physical_validity_feasible(charge_distribution_surface<Lyt>& cds_layout,
-                                  charge_distribution_surface<Lyt>& cds_canvas) const noexcept
-    {
-        auto min_energy = std::numeric_limits<double>::infinity();
-
-        uint64_t canvas_charge_index = 0;
-        cds_canvas.assign_charge_index(canvas_charge_index);
-
-        while (cds_canvas.get_charge_index_and_base().first <= cds_canvas.get_max_charge_index())
-        {
-            cds_canvas.foreach_cell(
-                [&](const auto& c)
-                {
-                    cds_layout.assign_charge_state(c, cds_canvas.get_charge_state(c),
-                                                   charge_index_mode::KEEP_CHARGE_INDEX);
-                });
-            cds_layout.update_after_charge_change(dependent_cell_mode::VARIABLE,
-                                                  energy_calculation::KEEP_OLD_ENERGY_VALUE);
-
-            if (cds_layout.is_physically_valid())
-            {
-                cds_layout.recompute_system_energy();
-                if (cds_layout.get_system_energy() + physical_constants::POP_STABILITY_ERR < min_energy)
-                {
-                    min_energy = cds_layout.get_system_energy();
-                }
-            }
-
-            if (cds_canvas.get_charge_index_and_base().first == cds_canvas.get_max_charge_index())
-            {
-                break;
-            }
-
-            canvas_charge_index++;
-            cds_canvas.assign_charge_index(canvas_charge_index);
-        }
-
-        if (min_energy < std::numeric_limits<double>::infinity())
-        {
-            return min_energy;
-        }
-
-        return std::nullopt;
-    }
-
-    /**
-     * This function iterates through various input patterns and output wire indices to determine if any configuration
-     * results in a physically valid layout with energy below the given energy value, indicating I/O signal instability.
-     *
-     * @param cds_layout The charge distribution surface layout to be modified and checked.
-     * @param cds_canvas The charge distribution of the canvas SiDBs.
-     * @param max_input_pattern_index The maximum index for input pattern
-     * @param input_pattern The specific input pattern for which the stability check is conducted.
-     * @param logical_correct_output_pattern The expected correct output pattern for the given input.
-     * @param minimal_energy_of_physically_valid_layout The minimum energy threshold below which the layout is
-     * considered unstable.
-     * @return `true` if the I/O signal is unstable, `false` otherwise.
-     */
-    [[nodiscard]] bool is_io_signal_unstable(charge_distribution_surface<Lyt>& cds_layout,
-                                             charge_distribution_surface<Lyt>& cds_canvas,
-                                             const uint64_t max_input_pattern_index, const uint64_t input_pattern,
-                                             const uint64_t logical_correct_output_pattern,
-                                             const double   minimal_energy_of_physically_valid_layout) const noexcept
-    {
-        for (auto kink_states_input = 0u; kink_states_input < max_input_pattern_index; ++kink_states_input)
-        {
-            for (auto output_wire_index = 0u; output_wire_index < std::pow(2, number_of_output_wires);
-                 output_wire_index++)
-            {
-                if (output_wire_index == logical_correct_output_pattern && kink_states_input == input_pattern)
-                {
-                    continue;
-                }
-
-                set_charge_distribution_of_input_wires_based_on_input_pattern(cds_layout, kink_states_input);
-                set_charge_distribution_of_output_wires_based_on_output_index(cds_layout, output_wire_index);
-
-                const auto physical_validity = is_physical_validity_feasible(cds_layout, cds_canvas);
-
-                if (physical_validity.has_value())
-                {
-                    if (physical_validity.value() + physical_constants::POP_STABILITY_ERR <
-                        minimal_energy_of_physically_valid_layout)
-                    {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * This function evaluates whether the given layout can be discarded since it cannot implement the given Boolean
-     * function. The pruning is subdivided into three single pruning steps: (1) discarding SiDB layouts with potentially
-     * positively charged SiDBs, (2) utilizing an efficient method to identify and discard SiDB layouts that do not
-     * satisfy physical model constraints under the I/O pin conditions required for the desired Boolean function, and
-     * (3) detecting I/O signal instability.
-     *
-     * @param current_layout The layout being evaluated for pruning.
-     * @param canvas_lyt The canvas layout comprising of the canvas SiDBs.
-     * @param dependent_cell A dependent-cell of the canvas SiDBs.
-     * @return `true` if the current layout can be pruned. `false` otherwise, which means that the layout is a candidate
-     * to be a valid gate implementation. Physical simulation is required as a second step to conduct the final
-     * validation.
-     */
-    [[nodiscard]] bool layout_can_be_pruned(const Lyt& current_layout, const Lyt& canvas_lyt,
-                                            const cell<Lyt>& dependent_cell) const noexcept
-    {
-        charge_distribution_surface<Lyt> cds_canvas{canvas_lyt, params.operational_params.simulation_parameters,
-                                                    sidb_charge_state::NEGATIVE,
-                                                    cds_configuration::CHARGE_LOCATION_ONLY};
-
-        cds_canvas.assign_dependent_cell(dependent_cell);
-
-        auto bii = bdl_input_iterator<Lyt>{current_layout, params.operational_params.input_bdl_iterator_params,
-                                           input_bdl_wires};
-
-        for (auto i = 0u; i < truth_table.front().num_bits(); ++i, ++bii)
-        {
-            charge_distribution_surface cds_layout{*bii, params.operational_params.simulation_parameters};
-
-            if (can_positive_charges_occur(cds_layout, params.operational_params.simulation_parameters))
-            {
-                return true;
-            }
-
-            cds_layout.assign_dependent_cell(dependent_cell);
-            set_charge_distribution_of_input_wires_based_on_input_pattern(cds_layout, i);
-            set_charge_distribution_of_output_wires_based_on_truth_table(cds_layout, i);
-
-            const auto physical_validity = is_physical_validity_feasible(cds_layout, cds_canvas);
-
-            if (physical_validity.has_value())
-            {
-                const auto output_index = determine_output_index_of_output(i);
-                if (is_io_signal_unstable(cds_layout, cds_canvas, truth_table.front().num_bits(), i, output_index,
-                                          physical_validity.value()))
-                {
-                    return true;
-                };
-            }
-            else
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
+    std::size_t number_of_threads{std::thread::hardware_concurrency()};
     /**
      * This function processes each layout to determine if it represents a valid gate implementation or if it can be
      * pruned by using three distinct physically-informed pruning steps. It leverages multi-threading to accelerate the
@@ -941,7 +615,7 @@ class design_sidb_gates_impl
      *
      * @return A vector containing the valid gate candidates that were not pruned.
      */
-    [[nodiscard]] std::vector<Lyt> run_pruning() const noexcept
+    [[nodiscard]] std::vector<Lyt> run_pruning() noexcept
     {
         std::vector<Lyt> gate_candidate = {};
 
@@ -958,30 +632,72 @@ class design_sidb_gates_impl
             auto current_layout = skeleton_layout.clone();
 
             cell<Lyt> dependent_cell{};
+
             canvas_lyt.foreach_cell(
-                [&](const auto& c)
+                [&current_layout, &dependent_cell](const auto& c)
                 {
                     current_layout.assign_cell_type(c, Lyt::technology::cell_type::LOGIC);
                     dependent_cell = c;
                 });
 
-            if (!layout_can_be_pruned(current_layout, canvas_lyt, dependent_cell))
+            charge_distribution_surface<Lyt> cds_canvas{canvas_lyt, params.operational_params.simulation_parameters,
+                                                        sidb_charge_state::NEGATIVE,
+                                                        cds_configuration::CHARGE_LOCATION_ONLY};
+
+            cds_canvas.assign_dependent_cell(dependent_cell);
+
+            auto bii = bdl_input_iterator<Lyt>{current_layout, params.operational_params.input_bdl_iterator_params,
+                                               input_bdl_wires};
+
+            detail::is_operational_impl<Lyt, TT> is_operational_impl{
+                current_layout, truth_table, params.operational_params, input_bdl_wires, output_bdl_wires, canvas_lyt};
+
+            for (auto i = 0u; i < truth_table.front().num_bits(); ++i, ++bii)
             {
-                const std::lock_guard lock{mutex_to_protect_gate_candidates};
-                gate_candidate.push_back(current_layout);
+                const auto reason_for_filtering =
+                    is_operational_impl.is_layout_invalid(bii.get_current_input_index(), cds_canvas);
+
+                if (reason_for_filtering.has_value())
+                {
+                    switch (reason_for_filtering.value())
+                    {
+                        case detail::layout_invalidity_reason::POTENTIAL_POSITIVE_CHARGES:
+                        {
+                            number_of_discarded_layouts_at_first_pruning++;
+                            break;
+                        }
+                        case detail::layout_invalidity_reason::PHYSICAL_INFEASIBILITY:
+                        {
+                            number_of_discarded_layouts_at_second_pruning++;
+                            break;
+                        }
+                        case detail::layout_invalidity_reason::IO_INSTABILITY:
+                        {
+                            number_of_discarded_layouts_at_third_pruning++;
+                            break;
+                        }
+                        default:
+                        {
+                            break;
+                        }
+                    }
+                    return;
+                }
             }
+
+            const std::lock_guard lock{mutex_to_protect_gate_candidates};
+            gate_candidate.push_back(current_layout);
         };
 
         gate_candidate.reserve(all_canvas_layouts.size());
 
-        const std::size_t number_of_threads =
-            std::min(static_cast<std::size_t>(std::thread::hardware_concurrency()), all_canvas_layouts.size());
-        const std::size_t chunk_size = (all_canvas_layouts.size() + number_of_threads - 1) / number_of_threads;
+        const std::size_t num_threads = std::min(number_of_threads, all_canvas_layouts.size());
+        const std::size_t chunk_size  = (all_canvas_layouts.size() + num_threads - 1) / num_threads;
 
         std::vector<std::thread> threads{};
-        threads.reserve(number_of_threads);
+        threads.reserve(num_threads);
 
-        for (std::size_t i = 0; i < number_of_threads; ++i)
+        for (std::size_t i = 0; i < num_threads; ++i)
         {
             threads.emplace_back(
                 [i, chunk_size, this, &conduct_pruning_steps]()
@@ -1006,7 +722,6 @@ class design_sidb_gates_impl
 
         return gate_candidate;
     }
-
     /**
      * This function calculates all combinations of distributing a given number of SiDBs across a specified number of
      * positions in the canvas. Each combination is then used to create a gate layout candidate.
@@ -1034,7 +749,6 @@ class design_sidb_gates_impl
 
         return designed_gate_layouts;
     }
-
     /**
      * This function adds SiDBs (given by indices) to the skeleton layout that is returned afterwards.
      *
@@ -1057,7 +771,6 @@ class design_sidb_gates_impl
 
         return lyt_copy;
     }
-
     /**
      * This function generates canvas SiDb layouts.
      *
@@ -1085,7 +798,7 @@ class design_sidb_gates_impl
 
         // the skeleton can already exhibit some canvas SiDBs (partially filled canvas)
         skeleton_layout.foreach_cell(
-            [&](const auto& c)
+            [this, &lyt](const auto& c)
             {
                 if (skeleton_layout.get_cell_type(c) == sidb_technology::cell_type::LOGIC)
                 {
@@ -1100,7 +813,7 @@ class design_sidb_gates_impl
      * This function makes sure that the underlying parameters for `is_operational` allow simulation results to be used
      * when the given parameter set indicates the use for it.
      *
-     * @param Parameters and settings for the gate designer.
+     * @param params Parameters and settings for the gate designer.
      * @return Parameters and settings for the gate designer for which the simulation results retention of the
      * underlying parameter for operational status assessment is set accordingly.
      */
@@ -1123,7 +836,8 @@ class design_sidb_gates_impl
     /**
      * Performs a sorting operation on the designed gate layouts, putting those in front for which the energetic gap
      * between the ground state and the first excited state is larger. For each designed gate layout, the minimum
-     * energetic gap is taken over each input.
+     * energetic gap is taken over each input. When the minima are equal for two designed gate layouts, the average
+     * energetic gap over each input is taken as a tiebreaker.
      *
      * @param designed_gate_layouts A vector of designed gate layouts to sort in place.
      * @param sim_results_per_input_for_each_gate_design The simulation results for each input of each designed gate
@@ -1137,7 +851,7 @@ class design_sidb_gates_impl
         std::vector<std::pair<Lyt, std::vector<std::vector<charge_distribution_surface<Lyt>>>>> pairs{};
         pairs.reserve(designed_gate_layouts.size());
 
-        for (auto i = 0; i < designed_gate_layouts.size(); ++i)
+        for (uint64_t i = 0; i < designed_gate_layouts.size(); ++i)
         {
             pairs.emplace_back(std::move(designed_gate_layouts.at(i)),
                                std::move(sim_results_per_input_for_each_gate_design.at(i)));
