@@ -23,8 +23,8 @@ namespace fiction
  * only layouts with up to 30 SiDBs can be simulated in a reasonable time. However, since all charge configurations are
  * checked for validity, 100 % simulation accuracy is guaranteed.
  *
- * @note This was the first exact simulation approach. However, it is replaced by *QuickExact* due to the much
- * better runtimes and more functionality.
+ * @note This was the first exact simulation approach. However, it is replaced by *QuickExact* and *ClusterComplete* due
+ * to the much better runtimes and more functionality.
  *
  * @tparam Lyt SiDB cell-level layout type.
  * @param lyt The layout to simulate.
