@@ -5,7 +5,7 @@
 #ifndef FICTION_IS_GROUND_STATE_HPP
 #define FICTION_IS_GROUND_STATE_HPP
 
-#include "fiction/algorithms/simulation/sidb/determine_groundstate_from_simulation_results.hpp"
+#include "fiction/algorithms/simulation/sidb/groundstate_from_simulation_result.hpp"
 #include "fiction/algorithms/simulation/sidb/sidb_simulation_result.hpp"
 #include "fiction/traits.hpp"
 
@@ -38,10 +38,9 @@ template <typename Lyt>
         return false;
     }
 
-    const auto ground_state_charge_distributions_exact = determine_groundstate_from_simulation_results(exact_results);
+    const auto ground_state_charge_distributions_exact = groundstate_from_simulation_result(exact_results);
 
-    const auto ground_state_charge_distributions_heuristic =
-        determine_groundstate_from_simulation_results(heuristic_results);
+    const auto ground_state_charge_distributions_heuristic = groundstate_from_simulation_result(heuristic_results);
 
     assert(ground_state_charge_distributions_heuristic.size() <= ground_state_charge_distributions_exact.size() &&
            "The heuristic results must be less equal than the exact results.");
