@@ -5,7 +5,9 @@
 #ifndef FICTION_WRITE_SVG_LAYOUT_HPP
 #define FICTION_WRITE_SVG_LAYOUT_HPP
 
+#include "fiction/layouts/bounding_box.hpp"
 #include "fiction/layouts/coordinates.hpp"
+#include "fiction/technology/sidb_charge_state.hpp"
 #include "fiction/traits.hpp"
 #include "utils/version_info.hpp"
 
@@ -13,9 +15,13 @@
 
 #include <array>
 #include <cmath>
+#include <cstdint>
 #include <exception>
+#include <fstream>
 #include <iostream>
+#include <optional>
 #include <sstream>
+#include <stdexcept>
 #include <string>
 #include <string_view>
 #include <tuple>
@@ -35,6 +41,61 @@ struct write_qca_layout_svg_params
      * Limit details to create smaller file sizes.
      */
     bool simple = false;
+};
+
+/**
+ * Parameters for writing SiDB layouts to SVG format.
+ */
+struct write_sidb_layout_svg_params
+{
+    /**
+     * Enumeration to specify the color mode for the SVG output.
+     */
+    enum class color_mode : uint8_t
+    {
+        /**
+         * Dark mode.
+         */
+        DARK,
+        /**
+         * Light mode.
+         */
+        LIGHT
+    };
+    /**
+     * Enumeration to specify if the H-Si lattice is plotted in addition to SiDBs.
+     */
+    enum class sidb_lattice_mode : uint8_t
+    {
+        /**
+         * Lattice is hidden. Only SiDBs are shown.
+         */
+        HIDE_LATTICE,
+        /**
+         * Lattice is shown.
+         */
+        SHOW_LATTICE
+    };
+    /**
+     * Size of the H-Si lattice points in SVG units.
+     */
+    double lattice_point_size = 0.3;
+    /**
+     * Size of the SiDB in SVG units.
+     */
+    double sidb_size = 0.9;
+    /**
+     * Border width of the SiDB.
+     */
+    double sidb_border_width = 0.3;
+    /**
+     * The color mode of the background for the SVG output.
+     */
+    color_mode color_background = color_mode::DARK;
+    /**
+     * The lattice mode of the SiDB layout.
+     */
+    sidb_lattice_mode lattice_mode = sidb_lattice_mode::SHOW_LATTICE;
 };
 
 template <typename Coordinate>
@@ -57,6 +118,56 @@ namespace detail
 
 namespace svg
 {
+
+// Color definitions
+inline constexpr const char* NEUTRAL_COLOR                   = "#6e7175";
+inline constexpr const char* NEGATIVE_COLOR                  = "#00ADAE";
+inline constexpr const char* POSITIVE_COLOR                  = "#E34857";
+inline constexpr const char* BACKGROUND_COLOR_BRIGHT         = "#ffffff";
+inline constexpr const char* BACKGROUND_COLOR_DARK           = "#25323D";
+inline constexpr const char* SIDB_DOT_DARK_MODE              = "#C8C8C8";
+inline constexpr const char* SIDB_DOT_BRIGHT_MODE            = "#ffffff";
+inline constexpr const char* SIDB_DOT_LINE_COLOR_DARK_MODE   = "#ffffff";
+inline constexpr const char* SIDB_DOT_LINE_COLOR_BRIGHT_MODE = "#C8C8C8";
+inline constexpr const char* SI_LATTICE                      = "#6e7175";
+
+// SVG Header with numbered placeholders
+inline constexpr const char* HEADER_TEMPLATE = R"(<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<!-- Generated by {0} ({1}) -->
+<svg
+    xmlns:dc="http://purl.org/dc/elements/1.1/"
+    xmlns:cc="http://creativecommons.org/ns#"
+    xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+    xmlns:svg="http://www.w3.org/2000/svg"
+    xmlns:xlink="http://www.w3.org/1999/xlink"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="{2} {3} {4} {5}"
+    version="1.1">
+    <metadata>
+        <rdf:RDF>
+            <cc:Work rdf:about="">
+                <dc:format>image/svg+xml</dc:format>
+                <dc:type rdf:resource="http://purl.org/dc/dcmitype/StillImage" />
+                <dc:title></dc:title>
+            </cc:Work>
+        </rdf:RDF>
+    </metadata>
+    {6} <!-- PATH_DEFINITION placeholder -->
+    {7} <!-- Background rectangle placeholder -->
+    <g>
+        {8} <!-- SVG content placeholder -->
+    </g>
+</svg>)";
+
+// SVG path definitions for lattice points and SiDBs
+static constexpr const char* PATH_DEFINITION_TEMPLATE = R"(
+<defs>
+    <!-- lattice point -->
+    <circle id="lattice_point" cx="0" cy="0" r="{0}"/>
+    <!-- SiDB -->
+    <circle id="sidb_color" cx="0" cy="0" r="{1}"/>
+</defs>
+)";
 
 // General layout
 inline constexpr const double VIEWBOX_DISTANCE = 11;
@@ -338,6 +449,269 @@ inline constexpr const char* SIMPLE_CELL = "<g\n"
 }  // namespace svg
 
 template <typename Lyt>
+class write_sidb_layout_svg_impl
+{
+  public:
+    /**
+     * Constructor.
+     */
+    write_sidb_layout_svg_impl(const Lyt& layout, std::ostream& stream, const write_sidb_layout_svg_params& p = {}) :
+            lyt{layout},
+            os{stream},
+            ps{p}
+    {
+        set_colors();
+    }
+
+    /**
+     * Runs the SVG generation process.
+     */
+    void run()
+    {
+        generate_svg();
+    }
+
+  private:
+    /**
+     * The SiDB layout to be written.
+     */
+    const Lyt& lyt;
+    /**
+     * The output stream to write into.
+     */
+    std::ostream& os;
+    /**
+     * Parameters for the SVG generation.
+     */
+    const write_sidb_layout_svg_params ps{};
+    /**
+     * The color mode for the SVG output.
+     */
+    std::string background_color;
+    /**
+     * The color of the SiDB without charge information.
+     */
+    std::string sidb_color;
+    /**
+     * The edge color of the SiDB without charge information.
+     */
+    std::string sidb_edge_color;
+    /**
+     * Sets the colors based on the color mode.
+     */
+    void set_colors() noexcept
+    {
+        if (ps.color_background == write_sidb_layout_svg_params::color_mode::LIGHT)
+        {
+            // fiction namespace is needed to avoid windows and linux conflict
+            background_color = fiction::detail::svg::BACKGROUND_COLOR_BRIGHT;
+            sidb_color       = fiction::detail::svg::SIDB_DOT_BRIGHT_MODE;
+            sidb_edge_color  = fiction::detail::svg::SIDB_DOT_LINE_COLOR_BRIGHT_MODE;
+        }
+        else
+        {
+            // fiction namespace is needed to avoid windows and linux conflict
+            background_color = fiction::detail::svg::BACKGROUND_COLOR_DARK;
+            sidb_color       = fiction::detail::svg::SIDB_DOT_DARK_MODE;
+            sidb_edge_color  = fiction::detail::svg::SIDB_DOT_LINE_COLOR_DARK_MODE;
+        }
+    }
+
+    /**
+     * Generates an SVG string representing an H-Si lattice point.
+     *
+     * @param x The x-coordinate of the lattice point.
+     * @param y The y-coordinate of the lattice point.
+     * @param fill_color The fill color of the lattice point.
+     *
+     * @return The SVG string representing the lattice point.
+     */
+    [[nodiscard]] std::string generate_lattice_point(const double x, const double y,
+                                                     const std::string& fill_color) const noexcept
+    {
+        return fmt::format(R"(<use xlink:href="#lattice_point" x="{0}" y="{1}" style="fill:{2};"/>)", x, y, fill_color);
+    }
+
+    /**
+     * Generates an SVG string representing an SiDB.
+     *
+     * @param x The x-coordinate of the SiDB.
+     * @param y The y-coordinate of the SiDB.
+     * @param charge_state The charge state of the SiDB.
+     *
+     * @return The SVG string representing the SiDB.
+     */
+    [[nodiscard]] std::string generate_sidb(const double x, const double y,
+                                            const std::optional<sidb_charge_state>& charge_state = std::nullopt) const
+    {
+        std::string fill_color   = sidb_color;
+        std::string border_color = sidb_edge_color;
+
+        auto fill_opacity = 1.0;
+
+        if (charge_state.has_value())
+        {
+            switch (charge_state.value())
+            {
+                case (sidb_charge_state::POSITIVE):
+                {
+                    fill_color   = fiction::detail::svg::POSITIVE_COLOR;
+                    border_color = fiction::detail::svg::POSITIVE_COLOR;
+                    break;
+                }
+                case (sidb_charge_state::NEGATIVE):
+                {
+                    fill_color   = fiction::detail::svg::NEGATIVE_COLOR;
+                    border_color = fiction::detail::svg::NEGATIVE_COLOR;
+                    break;
+                }
+                case (sidb_charge_state::NEUTRAL):
+                {
+                    fill_opacity = 0.0;
+                    break;
+                }
+                default:
+                {
+                    border_color = fiction::detail::svg::NEUTRAL_COLOR;
+                    fill_opacity = 0.0;
+                    break;
+                }
+            }
+        }
+
+        return fmt::format(
+            R"(<use xlink:href="#sidb_color" x="{0}" y="{1}" style="fill:{2}; fill-opacity:{3}; stroke:{4}; stroke-width:{5};"/>)",
+            x, y, fill_color, fill_opacity, border_color, ps.sidb_border_width);
+    }
+
+    /**
+     * Generates the SVG layout with both H-Si lattice points and SiDBs.
+     */
+    void generate_svg() const
+    {
+        std::stringstream svg_content{};
+
+        // Prepare the PATH_DEFINITION_TEMPLATE with the sizes
+        const std::string path_definition_template =
+            fmt::format(fiction::detail::svg::PATH_DEFINITION_TEMPLATE, ps.lattice_point_size, ps.sidb_size);
+
+        // Compute the bounding box of the layout
+        const auto bb        = bounding_box_2d{lyt};
+        const auto min_coord = bb.get_min();
+        const auto max_coord = bb.get_max();
+
+        if (ps.lattice_mode == write_sidb_layout_svg_params::sidb_lattice_mode::SHOW_LATTICE)
+        {
+            // Generate all lattice points
+            const auto all_coords = all_coordinates_in_spanned_area(min_coord, max_coord);
+
+            for (const auto& coord : all_coords)
+            {
+                // Shift coordinates for alignment
+                auto shifted_coord = coord;
+
+                shifted_coord.x += static_cast<decltype(shifted_coord.x)>(1);
+
+                if constexpr (has_siqad_coord_v<Lyt>)
+                {
+                    shifted_coord.y += static_cast<decltype(shifted_coord.y)>(1);
+                }
+                else
+                {
+                    shifted_coord.y += static_cast<decltype(shifted_coord.y)>(2);
+                }
+
+                const auto nm_pos = sidb_nm_position(lyt, shifted_coord);
+
+                svg_content << generate_lattice_point(nm_pos.first * 10, nm_pos.second * 10,
+                                                      fiction::detail::svg::SI_LATTICE);
+            }
+        }
+
+        std::vector<cell<Lyt>> all_cells{};
+        all_cells.reserve(lyt.num_cells());
+        // collect all cells
+        lyt.foreach_cell([&all_cells](const auto& cell) { all_cells.push_back(cell); });
+        std::sort(all_cells.begin(), all_cells.end());
+
+        for (const auto& cell : all_cells)
+        {
+            // Shift coordinates for alignment
+            auto shifted_cell = cell;
+
+            // shift for padding
+            shifted_cell.x += static_cast<decltype(shifted_cell.x)>(1);
+
+            if constexpr (has_siqad_coord_v<Lyt>)
+            {
+                shifted_cell.y += static_cast<decltype(shifted_cell.y)>(1);
+            }
+            else
+            {
+                shifted_cell.y += static_cast<decltype(shifted_cell.y)>(2);
+            }
+
+            const auto nm_pos = sidb_nm_position(lyt, shifted_cell);
+
+            if constexpr (is_charge_distribution_surface_v<Lyt>)
+            {
+                // If the layout has charge distribution information
+                const auto charge_state = lyt.get_charge_state(cell);
+                svg_content << generate_sidb(nm_pos.first * 10, nm_pos.second * 10, charge_state);
+            }
+            else
+            {
+                // Default SiDB dot without charge state
+                svg_content << generate_sidb(nm_pos.first * 10, nm_pos.second * 10);
+            }
+        };
+
+        auto shifted_max = max_coord;
+
+        // shift for padding
+
+        shifted_max.x += static_cast<decltype(shifted_max.x)>(2);
+
+        if constexpr (has_siqad_coord_v<Lyt>)
+        {
+            if (shifted_max.z == 1)
+            {
+                shifted_max.y += static_cast<decltype(shifted_max.y)>(2);
+                shifted_max.z = 0;
+            }
+            else
+            {
+                shifted_max.y += static_cast<decltype(shifted_max.y)>(1);
+                shifted_max.z = 1;
+            }
+        }
+
+        else
+        {
+            shifted_max.y += static_cast<decltype(shifted_max.y)>(3);
+        }
+
+        // Compute viewBox dimensions
+        const auto view_box_nm_min = sidb_nm_position(lyt, min_coord);
+        const auto view_box_nm_max = sidb_nm_position(lyt, shifted_max);
+
+        const auto viewbox_x      = view_box_nm_min.first * 10;
+        const auto viewbox_y      = view_box_nm_min.second * 10;
+        const auto viewbox_width  = (view_box_nm_max.first - view_box_nm_min.first) * 10;
+        const auto viewbox_height = (view_box_nm_max.second - view_box_nm_min.second) * 10;
+
+        // Generate background rectangle
+        const auto background_rect =
+            fmt::format(R"(<rect x="{0}" y="{1}" width="{2}" height="{3}" style="fill:{4};"/>)", viewbox_x, viewbox_y,
+                        viewbox_width, viewbox_height, background_color);
+
+        // Generate the final SVG content
+        os << fmt::format(fiction::detail::svg::HEADER_TEMPLATE, FICTION_VERSION, FICTION_REPO, viewbox_x, viewbox_y,
+                          viewbox_width, viewbox_height, path_definition_template, background_rect, svg_content.str());
+    }
+};
+
+template <typename Lyt>
 class write_qca_layout_svg_impl
 {
   public:
@@ -498,11 +872,11 @@ class write_qca_layout_svg_impl
                         if (lyt.is_synchronization_element(c))
                         {
                             cell_descriptions
-                                << fmt::format(desc_col.first, desc_col.second,
+                                << fmt::format(fmt::runtime(desc_col.first), desc_col.second,
                                                svg::STARTING_OFFSET_TILE_X + svg::STARTING_OFFSET_LATCH_CELL_X +
-                                                   c.x * svg::CELL_DISTANCE,
+                                                   (c.x * svg::CELL_DISTANCE),
                                                svg::STARTING_OFFSET_TILE_Y + svg::STARTING_OFFSET_LATCH_CELL_Y +
-                                                   c.y * svg::CELL_DISTANCE);
+                                                   (c.y * svg::CELL_DISTANCE));
 
                             is_sync_elem = true;
                         }
@@ -510,15 +884,15 @@ class write_qca_layout_svg_impl
                     if (!is_sync_elem)
                     {
                         cell_descriptions << fmt::format(
-                            desc_col.first, desc_col.second,
-                            svg::STARTING_OFFSET_TILE_X + svg::STARTING_OFFSET_CELL_X + c.x * svg::CELL_DISTANCE,
-                            svg::STARTING_OFFSET_TILE_Y + svg::STARTING_OFFSET_CELL_Y + c.y * svg::CELL_DISTANCE);
+                            fmt::runtime(desc_col.first), desc_col.second,
+                            svg::STARTING_OFFSET_TILE_X + svg::STARTING_OFFSET_CELL_X + (c.x * svg::CELL_DISTANCE),
+                            svg::STARTING_OFFSET_TILE_Y + svg::STARTING_OFFSET_CELL_Y + (c.y * svg::CELL_DISTANCE));
                     }
                 }
             });
 
-        const double viewbox_x = 2 * svg::VIEWBOX_DISTANCE + static_cast<double>(lyt.x() + 1) * svg::CELL_DISTANCE;
-        const double viewbox_y = 2 * svg::VIEWBOX_DISTANCE + static_cast<double>(lyt.y() + 1) * svg::CELL_DISTANCE;
+        const double viewbox_x = (2 * svg::VIEWBOX_DISTANCE) + (static_cast<double>(lyt.x() + 1) * svg::CELL_DISTANCE);
+        const double viewbox_y = (2 * svg::VIEWBOX_DISTANCE) + (static_cast<double>(lyt.y() + 1) * svg::CELL_DISTANCE);
 
         os << fmt::format(svg::HEADER, FICTION_VERSION, FICTION_REPO, viewbox_x, viewbox_y, cell_descriptions.str());
     }
@@ -606,9 +980,9 @@ class write_qca_layout_svg_impl
                         if (const auto latch_delay = lyt.get_synchronization_element(c); latch_delay > 0)
                         {
                             coord_to_latch_cells[tile_coords] = current_cells.append(
-                                fmt::format(desc_col.first, desc_col.second,
-                                            svg::STARTING_OFFSET_LATCH_CELL_X + in_tile.x * svg::CELL_DISTANCE,
-                                            svg::STARTING_OFFSET_LATCH_CELL_Y + in_tile.y * svg::CELL_DISTANCE));
+                                fmt::format(fmt::runtime(desc_col.first), desc_col.second,
+                                            svg::STARTING_OFFSET_LATCH_CELL_X + (in_tile.x * svg::CELL_DISTANCE),
+                                            svg::STARTING_OFFSET_LATCH_CELL_Y + (in_tile.y * svg::CELL_DISTANCE)));
 
                             is_sync_elem = true;
                         }
@@ -617,9 +991,9 @@ class write_qca_layout_svg_impl
                     if (!is_sync_elem)
                     {
                         coord_to_cells[tile_coords] = current_cells.append(
-                            fmt::format(desc_col.first, desc_col.second,
-                                        svg::STARTING_OFFSET_CELL_X + in_tile.x * svg::CELL_DISTANCE,
-                                        svg::STARTING_OFFSET_CELL_Y + in_tile.y * svg::CELL_DISTANCE));
+                            fmt::format(fmt::runtime(desc_col.first), desc_col.second,
+                                        svg::STARTING_OFFSET_CELL_X + (in_tile.x * svg::CELL_DISTANCE),
+                                        svg::STARTING_OFFSET_CELL_Y + (in_tile.y * svg::CELL_DISTANCE)));
                     }
                 }
             });
@@ -674,11 +1048,11 @@ class write_qca_layout_svg_impl
 
             const auto cell_descriptions = coord_to_cells[coord];
 
-            const double x_pos = svg::STARTING_OFFSET_TILE_X + coord.x * svg::TILE_DISTANCE;
-            const double y_pos = svg::STARTING_OFFSET_TILE_Y + coord.y * svg::TILE_DISTANCE;
+            const double x_pos = svg::STARTING_OFFSET_TILE_X + (coord.x * svg::TILE_DISTANCE);
+            const double y_pos = svg::STARTING_OFFSET_TILE_Y + (coord.y * svg::TILE_DISTANCE);
 
             const auto c_descr =
-                fmt::format(descr, x_pos, y_pos, tile_colors[czone], cell_descriptions,
+                fmt::format(fmt::runtime(descr), x_pos, y_pos, tile_colors[czone], cell_descriptions,
                             ps.simple ? "" : text_colors[czone], ps.simple ? "" : std::to_string(czone + 1));
 
             tile_descriptions << c_descr;
@@ -690,16 +1064,16 @@ class write_qca_layout_svg_impl
             for (const auto& [coord, ldscr] : coord_to_latch_tile)
             {
                 const auto [descr, czone_up, latch_delay] = ldscr;
-                const auto czone_lo                       = czone_up + latch_delay % lyt.num_clocks();
+                const auto czone_lo                       = czone_up + (latch_delay % lyt.num_clocks());
 
                 const auto cell_descriptions = coord_to_latch_cells[coord];
 
-                const double x_pos = svg::STARTING_OFFSET_LATCH_X + coord.x * svg::TILE_DISTANCE;
-                const double y_pos = svg::STARTING_OFFSET_LATCH_Y + coord.y * svg::TILE_DISTANCE;
+                const double x_pos = svg::STARTING_OFFSET_LATCH_X + (coord.x * svg::TILE_DISTANCE);
+                const double y_pos = svg::STARTING_OFFSET_LATCH_Y + (coord.y * svg::TILE_DISTANCE);
 
                 const auto t_descr =
-                    fmt::format(descr, x_pos, y_pos, tile_colors[czone_lo], tile_colors[czone_up], cell_descriptions,
-                                text_colors[czone_up], ps.simple ? "" : std::to_string(czone_up + 1),
+                    fmt::format(fmt::runtime(descr), x_pos, y_pos, tile_colors[czone_lo], tile_colors[czone_up],
+                                cell_descriptions, text_colors[czone_up], ps.simple ? "" : std::to_string(czone_up + 1),
                                 text_colors[czone_lo], ps.simple ? "" : std::to_string(czone_lo + 1));
 
                 tile_descriptions << t_descr;
@@ -708,8 +1082,8 @@ class write_qca_layout_svg_impl
 
         const coordinate<Lyt> length = {(lyt.x() + 1) / lyt.get_tile_size_x(), (lyt.y() + 1) / lyt.get_tile_size_y()};
 
-        const double viewbox_x = 2 * svg::VIEWBOX_DISTANCE + length.x * svg::TILE_DISTANCE;
-        const double viewbox_y = 2 * svg::VIEWBOX_DISTANCE + length.y * svg::TILE_DISTANCE;
+        const double viewbox_x = (2 * svg::VIEWBOX_DISTANCE) + (length.x * svg::TILE_DISTANCE);
+        const double viewbox_y = (2 * svg::VIEWBOX_DISTANCE) + (length.y * svg::TILE_DISTANCE);
 
         os << fmt::format(svg::HEADER, FICTION_VERSION, FICTION_REPO, viewbox_x, viewbox_y, tile_descriptions.str());
     }
@@ -768,6 +1142,56 @@ void write_qca_layout_svg(const Lyt& lyt, const std::string_view& filename, cons
     }
 
     write_qca_layout_svg(lyt, os, ps);
+    os.close();
+}
+
+/**
+ * Writes an SVG representation of an SiDB cell-level SiDB layout into an output stream.
+ *
+ * @note SiDB defects are not supported yet.
+ *
+ * @tparam Lyt SiDB cell-level layout type.
+ * @param lyt The layout to be written.
+ * @param os The output stream to write into.
+ * @param ps Parameters.
+ */
+template <typename Lyt>
+void write_sidb_layout_svg(const Lyt& lyt, std::ostream& os, const write_sidb_layout_svg_params& ps = {})
+{
+    static_assert(is_cell_level_layout_v<Lyt>, "Lyt is not a cell-level layout");
+    static_assert(has_sidb_technology_v<Lyt>, "Lyt must be a SiDB layout");
+    static_assert(!is_sidb_defect_surface_v<Lyt>, "SiDB defects are not supported");
+
+    detail::write_sidb_layout_svg_impl<Lyt> p{lyt, os, ps};
+    p.run();
+}
+
+/**
+ * Writes an SVG representation of an SiDB cell-level SiDB layout into a file.
+ *
+ * @note SiDB defects are not supported yet.
+ *
+ * @tparam Lyt SiDB cell-level layout type.
+ * @param lyt The layout to be written.
+ * @param filename The file name to create and write into.
+ * @param ps Parameters.
+ */
+template <typename Lyt>
+void write_sidb_layout_svg(const Lyt& lyt, const std::string_view& filename,
+                           const write_sidb_layout_svg_params& ps = {})
+{
+    static_assert(is_cell_level_layout_v<Lyt>, "Lyt is not a cell-level layout");
+    static_assert(has_sidb_technology_v<Lyt>, "Lyt must be a SiDB layout");
+    static_assert(!is_sidb_defect_surface_v<Lyt>, "SiDB defects are not supported");
+
+    std::ofstream os{filename.data(), std::ofstream::out};
+
+    if (!os.is_open())
+    {
+        throw std::ofstream::failure("Could not open file");
+    }
+
+    write_sidb_layout_svg(lyt, os, ps);
     os.close();
 }
 
