@@ -18,7 +18,6 @@
 #include <optional>
 #include <string>
 #include <string_view>
-#include <type_traits>
 #include <utility>
 
 namespace fiction
@@ -168,16 +167,24 @@ inline constexpr const char* RES           = "RES";
 inline constexpr const char* ESR           = "ESR";
 inline constexpr const char* CFE           = "CFE";
 inline constexpr const char* RIPPLE        = "RIPPLE";
+inline constexpr const char* SRS           = "SRS";
 inline constexpr const char* BANCS         = "BANCS";
 
 }  // namespace clock_name
 
-enum class num_clks
+enum class num_clks : uint8_t
 {
+    /**
+     * Three clocks.
+     */
     THREE,
+    /**
+     * Four clocks.
+     */
     FOUR
 };
-
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wsign-conversion"
 /**
  * Returns an irregular clocking that maps every coordinate to the standard clock. It is intended to be overridden.
  *
@@ -675,6 +682,35 @@ static auto ripple_clocking() noexcept
     // clang-format on
 }
 /**
+ * Returns the SRS clocking as defined in \"Simple, robust and systematic QCA clocking scheme for area-efficient
+ * nanocircuits\" by Mrinal Goswami, Tonmoy Jyoti Sharma, and Arpita Nath Boruah in International Journal of Electronics
+ * Letters 2025.
+ *
+ * @tparam Lyt Clocked layout type.
+ * @return SRS clocking scheme.
+ */
+template <typename Lyt>
+static auto srs_clocking() noexcept
+{
+    // clang-format off
+
+    static const typename clocking_scheme<clock_zone<Lyt>>::clock_function srs_clock_function =
+        [](const clock_zone<Lyt>& cz) noexcept
+    {
+        static constexpr std::array<std::array<typename clocking_scheme<clock_zone<Lyt>>::clock_number, 4u>, 4u> cutout{
+            {{{1, 0, 3, 2}},
+             {{2, 3, 0, 1}},
+             {{3, 2, 3, 0}},
+             {{0, 1, 2, 1}}}};
+
+        return cutout[cz.y % 4ul][cz.x % 4ul];
+    };
+
+    return clocking_scheme{clock_name::SRS, srs_clock_function, std::min(Lyt::max_fanin_size, 3u), 3u, 4u, true};
+
+    // clang-format on
+}
+/**
  * Returns the BANCS clocking as defined in \"BANCS: Bidirectional Alternating Nanomagnetic Clocking Scheme\" by
  * Ruan Evangelista Formigoni, Omar P. Vilela Neto, and Jose Augusto M. Nacif in SBCCI 2018.
  *
@@ -704,6 +740,7 @@ static auto bancs_clocking() noexcept
 
     // clang-format on
 }
+#pragma GCC diagnostic pop
 /**
  * Returns a smart pointer to the given scheme.
  *
@@ -782,6 +819,7 @@ std::optional<clocking_scheme<clock_zone<Lyt>>> get_clocking_scheme(const std::s
         {clock_name::ESR, esr_clocking<Lyt>()},
         {clock_name::CFE, cfe_clocking<Lyt>()},
         {clock_name::RIPPLE, ripple_clocking<Lyt>()},
+        {clock_name::SRS, srs_clocking<Lyt>()},
         {clock_name::BANCS, bancs_clocking<Lyt>()}};
 
     std::string upper_name = name.data();
