@@ -1082,7 +1082,7 @@ struct formatter<fiction::offset::ucoord_t>
     template <typename FormatContext>
     auto format(const fiction::offset::ucoord_t& c, FormatContext& ctx) const
     {
-        return format_to(ctx.out(), "({},{},{})", c.x, c.y, c.z);
+        return format_to(ctx.out(), runtime("({},{},{})"), c.x, c.y, c.z);
     }
 };
 // make cube::coord_t compatible with fmt::format
@@ -1098,7 +1098,7 @@ struct formatter<fiction::cube::coord_t>
     template <typename FormatContext>
     auto format(const fiction::cube::coord_t& c, FormatContext& ctx) const
     {
-        return format_to(ctx.out(), "({},{},{})", c.x, c.y, c.z);
+        return format_to(ctx.out(), runtime("({},{},{})"), c.x, c.y, c.z);
     }
 };
 // make siqad::coord_t compatible with fmt::format
@@ -1114,7 +1114,7 @@ struct formatter<fiction::siqad::coord_t>
     template <typename FormatContext>
     auto format(const fiction::siqad::coord_t& c, FormatContext& ctx) const
     {
-        return format_to(ctx.out(), "({},{},{})", c.x, c.y, c.z);
+        return format_to(ctx.out(), runtime("({},{},{})"), c.x, c.y, c.z);
     }
 };
 
