@@ -141,7 +141,7 @@ enum class charge_index_mode : uint8_t
     /**
      * The charge state is assigned to the cell but the old charge index is kept.
      */
-    KEEP_CHARGE_INDEX,
+    KEEP_CHARGE_INDEX
 };
 
 /**
@@ -149,7 +149,7 @@ enum class charge_index_mode : uint8_t
  * SiDBs' charge states.
  *
  * @tparam Lyt SiDB cell-level layout type.
- * @tparam has_sidb_charge_distribution Automatically determines whether a charge distribution interface is already
+ * @tparam has_charge_distribution_interface Automatically determines whether a charge distribution interface is already
  * present.
  */
 template <typename Lyt, bool has_charge_distribution_interface =
@@ -174,7 +174,7 @@ class charge_distribution_surface<Lyt, false> : public Lyt
     {
       private:
         /**
-         * The distance matrix is a vector of vectors storing the euclidean distance in nm.
+         * The distance matrix is a vector of vectors storing the Euclidean distance in nm.
          */
         using distance_matrix = std::vector<std::vector<double>>;
         /**
@@ -360,6 +360,17 @@ class charge_distribution_surface<Lyt, false> : public Lyt
         return *this;
     }
     /**
+     * Clones the current charge distribution surface and returns a deep copy.
+     *
+     * @return A deep copy of the current charge_distribution_surface, preserving all its properties.
+     */
+    [[nodiscard]] charge_distribution_surface clone() const noexcept
+    {
+        charge_distribution_surface copy{Lyt::clone()};
+        copy.strg = std::make_shared<charge_distribution_storage>(*strg);
+        return copy;
+    }
+    /**
      * This function returns the locations of all SiDBs in nm of the form `(x,y)`.
      *
      * @return Vector of SiDB nanometer positions (unit: nm).
@@ -371,8 +382,8 @@ class charge_distribution_surface<Lyt, false> : public Lyt
 
         for (const auto& c : strg->sidb_order)
         {
-            auto pos = sidb_nm_position<Lyt>(Lyt{}, c);
-            positions.push_back(std::make_pair(pos.first, pos.second));
+            auto pos = sidb_nm_position<Lyt>(*this, c);
+            positions.emplace_back(pos.first, pos.second);
         }
 
         return positions;
@@ -901,6 +912,17 @@ class charge_distribution_surface<Lyt, false> : public Lyt
         return std::nullopt;
     }
     /**
+     * This function allows the local electrostatic potential for some given index position to be set externally.
+     *
+     * @param index The index defining the SiDB position.
+     * @param loc_pot The local electrostatic potential to assign to the given index position (unit: V).
+     */
+    void assign_local_potential_by_index(const uint64_t index, const double loc_pot) noexcept
+    {
+        assert(index < strg->local_pot.size());
+        strg->local_pot[index] = loc_pot;
+    }
+    /**
      * This function assign the electrostatic system energy to zero (unit: eV). It can be used if only one SiDB is
      * charged.
      */
@@ -978,7 +1000,42 @@ class charge_distribution_surface<Lyt, false> : public Lyt
         this->validity_check();
     }
     /**
-     * The physically validity of the current charge distribution is evaluated and stored in the storage struct. A
+     * The configuration stability of the current charge distribution is evaluated. It is performed as the last check
+     * towards a judgement of physical validity of the present charge distribution layout.
+     *
+     * @return `true` if and only if the present charge distribution layout is deemed to be configuration stable.
+     */
+    [[nodiscard]] bool is_configuration_stable() const noexcept
+    {
+        const auto hop_del =
+            [this](const uint64_t c1, const uint64_t c2)  // energy change when charge hops between two SiDBs.
+        { return strg->local_pot[c1] - strg->local_pot[c2] - strg->pot_mat[c1][c2]; };
+
+        for (uint64_t i = 0u; i < strg->local_pot.size(); ++i)
+        {
+            if (strg->cell_charge[i] == sidb_charge_state::POSITIVE)  // we do nothing with SiDB+
+            {
+                continue;
+            }
+
+            for (uint64_t j = 0u; j < strg->local_pot.size(); j++)
+            {
+                if (const auto e_del = hop_del(i, j);
+                    (charge_state_to_sign(strg->cell_charge[j]) > charge_state_to_sign(strg->cell_charge[i])) &&
+                    (e_del < -physical_constants::POP_STABILITY_ERR))  // Checks if energetically favored
+                                                                       // hops exist between two SiDBs.
+                {
+                    return false;
+                }
+            }
+        }
+
+        // If there is no jump that leads to a decrease in the potential energy of the system, the given charge
+        // distribution satisfies metastability.
+        return true;
+    }
+    /**
+     * The physical validity of the current charge distribution is evaluated and stored in the storage struct. A
      * charge distribution is valid if the *Population Stability* and the *Configuration Stability* is fulfilled.
      */
     void validity_check() noexcept
@@ -987,15 +1044,15 @@ class charge_distribution_surface<Lyt, false> : public Lyt
         uint64_t   for_loop_counter                           = 0;
         const auto mu_p                                       = strg->simulation_parameters.mu_plus();
 
-        for (const auto& it : strg->local_pot)  // this for-loop checks if the "population stability" is fulfilled.
+        for (const auto& v : strg->local_pot)  // this for-loop checks if the "population stability" is fulfilled.
         {
             bool valid = (((strg->cell_charge[for_loop_counter] == sidb_charge_state::NEGATIVE) &&
-                           (-it + strg->simulation_parameters.mu_minus < physical_constants::POP_STABILITY_ERR)) ||
+                           (-v + strg->simulation_parameters.mu_minus < physical_constants::POP_STABILITY_ERR)) ||
                           ((strg->cell_charge[for_loop_counter] == sidb_charge_state::POSITIVE) &&
-                           (-it + mu_p > -physical_constants::POP_STABILITY_ERR)) ||
+                           (-v + mu_p > -physical_constants::POP_STABILITY_ERR)) ||
                           ((strg->cell_charge[for_loop_counter] == sidb_charge_state::NEUTRAL) &&
-                           (-it + strg->simulation_parameters.mu_minus > physical_constants::POP_STABILITY_ERR) &&
-                           (-it + mu_p < physical_constants::POP_STABILITY_ERR)));
+                           (-v + strg->simulation_parameters.mu_minus > -physical_constants::POP_STABILITY_ERR) &&
+                           (-v + mu_p < physical_constants::POP_STABILITY_ERR)));
             for_loop_counter += 1;
             if (!valid)
             {
@@ -1010,40 +1067,7 @@ class charge_distribution_surface<Lyt, false> : public Lyt
             (for_loop_counter >
              0))  // if population stability is fulfilled for all SiDBs, the "configuration stability" is checked.
         {
-            const auto hop_del =
-                [this](const uint64_t c1, const uint64_t c2)  // energy change when charge hops between two SiDBs.
-            { return strg->local_pot[c1] - strg->local_pot[c2] - strg->pot_mat[c1][c2]; };
-
-            uint64_t hop_counter = 0;
-            for (uint64_t i = 0u; i < strg->local_pot.size(); ++i)
-            {
-                if (strg->cell_charge[i] == sidb_charge_state::POSITIVE)  // we do nothing with SiDB+
-                {
-                    continue;
-                }
-
-                for (uint64_t j = 0u; j < strg->local_pot.size(); j++)
-                {
-                    if (hop_counter == 1)
-                    {
-                        break;
-                    }
-
-                    if (const auto e_del = hop_del(i, j);
-                        (charge_state_to_sign(strg->cell_charge[j]) > charge_state_to_sign(strg->cell_charge[i])) &&
-                        (e_del < -physical_constants::POP_STABILITY_ERR))  // Checks if energetically favored
-                                                                           // hops exist between two SiDBs.
-                    {
-                        hop_counter = 1;
-
-                        break;
-                    }
-                }
-            }
-
-            // If there is no jump that leads to a decrease in the potential energy of the system, the given charge
-            // distribution satisfies metastability.
-            strg->validity = hop_counter == 0;
+            strg->validity = is_configuration_stable();
         }
     }
     /**
@@ -1055,7 +1079,13 @@ class charge_distribution_surface<Lyt, false> : public Lyt
     {
         return strg->validity;
     }
-
+    /**
+     * This function declares present charge distribution layout as physically valid by external judgement.
+     */
+    void declare_physically_valid() noexcept
+    {
+        strg->validity = true;
+    }
     /**
      * The charge distribution of the charge distribution surface is converted to a unique index. It is used to map
      * every possible charge distribution of an SiDB layout to a unique index.
@@ -1315,7 +1345,7 @@ class charge_distribution_surface<Lyt, false> : public Lyt
     }
     /**
      * This function can be used to assign a global external electrostatic potential in Volt (unit: V) to the layout
-     * (e.g this could be a planar external electrode).
+     * (e.g this could be a planar external electrode). It is added to previously stored values.
      *
      * @param potential_value Value of the global external electrostatic potential in Volt (e.g. -0.3).
      * Charge-transition levels are shifted by this value.
@@ -1328,7 +1358,7 @@ class charge_distribution_surface<Lyt, false> : public Lyt
         if (potential_value != 0.0)
         {
             this->foreach_cell([this, &potential_value](const auto& c)
-                               { strg->local_external_pot.insert({c, potential_value}); });
+                               { strg->local_external_pot[c] += potential_value; });
             this->update_after_charge_change(dep_cell);
         }
     }
@@ -1412,9 +1442,10 @@ class charge_distribution_surface<Lyt, false> : public Lyt
      */
     [[nodiscard]] std::size_t num_negative_sidbs() const noexcept
     {
-        uint64_t count_negative_sidbs = 0;
+        uint64_t count_negative_sidbs = 0u;
         this->foreach_cell([this, &count_negative_sidbs](const auto& c)
-                           { count_negative_sidbs += (get_charge_state(c) == sidb_charge_state::NEGATIVE) ? 1 : 0; });
+                           { count_negative_sidbs += (get_charge_state(c) == sidb_charge_state::NEGATIVE) ? 1u : 0u; });
+
         return count_negative_sidbs;
     }
     /**
@@ -1426,7 +1457,8 @@ class charge_distribution_surface<Lyt, false> : public Lyt
     {
         std::size_t count_neutral_sidbs = 0;
         this->foreach_cell([this, &count_neutral_sidbs](const auto& c)
-                           { count_neutral_sidbs += (get_charge_state(c) == sidb_charge_state::NEUTRAL) ? 1 : 0; });
+                           { count_neutral_sidbs += (get_charge_state(c) == sidb_charge_state::NEUTRAL) ? 1u : 0u; });
+
         return count_neutral_sidbs;
     }
     /**
@@ -1436,9 +1468,10 @@ class charge_distribution_surface<Lyt, false> : public Lyt
      */
     [[nodiscard]] std::size_t num_positive_sidbs() const noexcept
     {
-        std::size_t count_positive_sidbs = 0;
+        std::size_t count_positive_sidbs = 0u;
         this->foreach_cell([this, &count_positive_sidbs](const auto& c)
-                           { count_positive_sidbs += (get_charge_state(c) == sidb_charge_state::POSITIVE) ? 1 : 0; });
+                           { count_positive_sidbs += (get_charge_state(c) == sidb_charge_state::POSITIVE) ? 1u : 0u; });
+
         return count_positive_sidbs;
     }
     /**
@@ -1573,8 +1606,8 @@ class charge_distribution_surface<Lyt, false> : public Lyt
                (distance * 1e-9) * std::exp(-distance / defect.lambda_tf) * physical_constants::ELEMENTARY_CHARGE;
     }
     /**
-     * This function can be used to assign an external local electrostatic potential in Volt to the layout. All
-     * important attributes of the charge layout are updated automatically.
+     * This function can be used to assign an external local electrostatic potential in Volt to the layout, which is
+     * added to previously stored values. All important attributes of the charge layout are updated automatically.
      *
      * @param cell Cell to which the local external potential is applied.
      * @param external_voltage External electrostatic potential in Volt applied to different cells.
@@ -1582,8 +1615,15 @@ class charge_distribution_surface<Lyt, false> : public Lyt
     void
     assign_local_external_potential(const std::unordered_map<typename Lyt::cell, double>& external_potential) noexcept
     {
-        strg->local_external_pot = external_potential;
-        this->update_after_charge_change();
+        for (const auto& [c, pot] : external_potential)
+        {
+            strg->local_external_pot[c] += external_potential.at(c);
+        }
+
+        if (!external_potential.empty())
+        {
+            this->update_after_charge_change();
+        }
     }
     /**
      * This function returns the local external electrostatic potential in Volt applied to the layout.
@@ -1591,16 +1631,25 @@ class charge_distribution_surface<Lyt, false> : public Lyt
      * @return External electrostatic potential as unordered map. The cell is used as key and the external
      * electrostatic potential in Volt (unit: V) at its position as value.
      */
-    std::unordered_map<typename Lyt::cell, double> get_local_external_potentials() noexcept
+    std::unordered_map<typename Lyt::cell, double> get_local_external_potentials() const noexcept
     {
         return strg->local_external_pot;
+    }
+    /**
+     * This function can be used to reset all external local electrostatic potentials to 0 Volt. All important
+     * attributes of the charge layout are updated automatically.
+     */
+    void reset_local_external_potentials() noexcept
+    {
+        strg->local_external_pot.clear();
+        this->update_after_charge_change();
     }
     /**
      * This function returns the local electrostatic potentials which are generated by defects.
      *
      * @return Local electrostatic potential in Volt generated by the defects at each each cell.
      */
-    std::unordered_map<typename Lyt::cell, double> get_local_defect_potentials() noexcept
+    std::unordered_map<typename Lyt::cell, double> get_local_defect_potentials() const noexcept
     {
         return strg->defect_local_pot;
     }
@@ -1609,7 +1658,7 @@ class charge_distribution_surface<Lyt, false> : public Lyt
      *
      * @return Placed defects with cell position and type.
      */
-    std::unordered_map<typename Lyt::cell, const sidb_defect> get_defects() noexcept
+    std::unordered_map<typename Lyt::cell, const sidb_defect> get_defects() const noexcept
     {
         return strg->defects;
     }
@@ -1757,7 +1806,7 @@ class charge_distribution_surface<Lyt, false> : public Lyt
      * @param engine The simulation engine used.
      */
     void increase_charge_index_of_sub_layout_by_one(
-        const dependent_cell_mode          dependent_cell_fixed    = dependent_cell_mode::FIXED,
+        const dependent_cell_mode          dependent_cell          = dependent_cell_mode::FIXED,
         const energy_calculation           recompute_system_energy = energy_calculation::UPDATE_ENERGY,
         const charge_distribution_history  consider_history        = charge_distribution_history::NEGLECT,
         const exact_sidb_simulation_engine engine                  = exact_sidb_simulation_engine::QUICKEXACT) noexcept
@@ -1773,7 +1822,7 @@ class charge_distribution_surface<Lyt, false> : public Lyt
             {
                 this->index_to_charge_distribution(charge_index_recomputation::IGNORE_LEADING_ZEROES);
             }
-            this->update_after_charge_change(dependent_cell_fixed, recompute_system_energy, consider_history);
+            this->update_after_charge_change(dependent_cell, recompute_system_energy, consider_history);
         }
     }
     /**
