@@ -8,9 +8,10 @@
 #include "utils/blueprints/layout_blueprints.hpp"
 
 #include <fiction/algorithms/simulation/sidb/critical_temperature.hpp>
+#include <fiction/algorithms/simulation/sidb/is_operational.hpp>
+#include <fiction/algorithms/simulation/sidb/sidb_simulation_engine.hpp>
 #include <fiction/algorithms/simulation/sidb/sidb_simulation_parameters.hpp>
 #include <fiction/technology/cell_technologies.hpp>
-#include <fiction/technology/charge_distribution_surface.hpp>
 #include <fiction/types.hpp>
 #include <fiction/utils/truth_table_utils.hpp>
 
@@ -37,12 +38,12 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
         lyt.assign_cell_type({6, 1, 0}, sidb_technology::cell_type::OUTPUT);
         lyt.assign_cell_type({8, 1, 0}, sidb_technology::cell_type::OUTPUT);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::APPROXIMATE;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 0;
-        params.alpha                 = 0.0;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKSIM;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 0;
+        params.alpha                                    = 0.0;
 
         const auto ct =
             critical_temperature_gate_based<TestType>(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
@@ -53,17 +54,30 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
     SECTION("No SiDB")
     {
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_gate_based<TestType>(lyt, std::vector<tt>{tt{}}, params, &critical_stats);
+        const auto ct_qe =
+            critical_temperature_gate_based<TestType>(lyt, std::vector<tt>{tt{}}, params, &critical_stats);
 
         CHECK(critical_stats.num_valid_lyt == 0);
-        CHECK(ct == 0.0);
+        CHECK(ct_qe == 0.0);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based<TestType>(lyt, std::vector<tt>{tt{}}, params, &critical_stats);
+
+        CHECK(critical_stats.num_valid_lyt == 0);
+        CHECK(ct_cc == 0.0);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("Not working diagonal wire where positively charged SiDBs can occur")
@@ -89,16 +103,27 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
         lyt.assign_cell_type({36, 19, 0}, sidb_technology::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
         const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
 
         CHECK(ct == 0.0);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
+
+        CHECK(ct_cc == 0.0);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("four SiDBs with two valid charge distributions, QuickExact")
@@ -109,19 +134,32 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
         lyt.assign_cell_type({2, 0}, TestType::cell_type::NORMAL);
         lyt.assign_cell_type({2, 2}, TestType::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_non_gate_based(lyt, params, &critical_stats);
+        const auto ct_qe = critical_temperature_non_gate_based(lyt, params, &critical_stats);
 
         CHECK(critical_stats.num_valid_lyt == 2);
         CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
                    Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
-        CHECK(ct == 350);
+        CHECK(ct_qe == 350);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc = critical_temperature_non_gate_based(lyt, params, &critical_stats);
+
+        CHECK(critical_stats.num_valid_lyt == 2);
+        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                   Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
+        CHECK(ct_cc == 350);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("Y-shaped SiQAD AND gate")
@@ -145,18 +183,32 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
         sim_params.mu_minus = -0.28;
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
 
         CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
                    Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
-        CHECK(ct == 350);
+        CHECK(ct_qe == 350);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+
+        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                   Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
+        CHECK(ct_cc == 350);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("Y-shaped SiQAD OR gate")
@@ -165,18 +217,31 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
         sim_params.mu_minus = -0.28;
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.input_iterator_params.input_bdl_config =
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.operational_params.input_bdl_iterator_params.input_bdl_config =
             bdl_input_iterator_params::input_bdl_configuration::PERTURBER_ABSENCE_ENCODED;
-        params.input_iterator_params.bdl_wire_params.threshold_bdl_interdistance = 1.5;
+        params.operational_params.input_bdl_iterator_params.bdl_wire_params.threshold_bdl_interdistance = 1.5;
 
-        const auto ct =
+        const auto ct_qe =
             critical_temperature_gate_based(lyt_or_gate, std::vector<tt>{create_or_tt()}, params, &critical_stats);
 
         CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
                    Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
-        CHECK(ct == 400);
+        CHECK(ct_qe == 400);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt_or_gate, std::vector<tt>{create_or_tt()}, params, &critical_stats);
+
+        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                   Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
+        CHECK(ct_cc == 400);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("Bestagon AND gate, QuickExact")
@@ -210,18 +275,59 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
         lyt.assign_cell_type({36, 19, 0}, sidb_technology::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+        SECTION("Kinks are allowed")
+        {
+            const auto ct_qe =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
 
-        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
-                   Catch::Matchers::WithinAbs(26.02, 0.01));
-        CHECK_THAT(std::abs(ct - 59.19), Catch::Matchers::WithinAbs(0.00, 0.01));
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                       Catch::Matchers::WithinAbs(26.02, 0.01));
+            CHECK_THAT(std::abs(ct_qe - 59.19), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+            params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+            const auto ct_cc =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                       Catch::Matchers::WithinAbs(26.02, 0.01));
+            CHECK_THAT(std::abs(ct_cc - 59.19), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+        }
+        SECTION("Kinks are not allowed")
+        {
+            params.operational_params.op_condition = is_operational_params::operational_condition::REJECT_KINKS;
+            params.operational_params.input_bdl_iterator_params.bdl_wire_params.threshold_bdl_interdistance = 2.5;
+            const auto ct_qe =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                       Catch::Matchers::WithinAbs(5.1153718076, 0.01));
+            CHECK_THAT(std::abs(ct_qe - 13.36), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+            params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+            const auto ct_cc =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                       Catch::Matchers::WithinAbs(5.1153718076, 0.01));
+            CHECK_THAT(std::abs(ct_cc - 13.36), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+        }
     }
 
     SECTION("Bestagon AND gate, QuickSim")
@@ -255,16 +361,28 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
         lyt.assign_cell_type({36, 19, 0}, sidb_technology::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::APPROXIMATE;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 500;
-        params.alpha                 = 0.6;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKSIM;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 500;
+        params.alpha                                    = 0.6;
 
-        const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+        const auto ct_qs =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
 
-        CHECK(ct > 0);
+        CHECK(ct_qs > 0);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+
+        CHECK(ct_cc > 0);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("Bestagon FO2 gate")
@@ -297,22 +415,92 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
         lyt.assign_cell_type({36, 19, 0}, sidb_technology::cell_type::NORMAL);
         lyt.assign_cell_type({2, 19, 0}, sidb_technology::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct =
-            critical_temperature_gate_based(lyt, std::vector<tt>{create_fan_out_tt()}, params, &critical_stats);
+        SECTION("Kinks are allowed")
+        {
+            const auto ct_qe =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_fan_out_tt()}, params, &critical_stats);
 
-        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.56),
-                   Catch::Matchers::WithinAbs(0.00, 0.01));
-        CHECK_THAT(std::abs(ct - 1.46), Catch::Matchers::WithinAbs(0.00, 0.01));
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.56),
+                       Catch::Matchers::WithinAbs(0.00, 0.01));
+            CHECK_THAT(std::abs(ct_qe - 1.46), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+            params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+            const auto ct_cc =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_fan_out_tt()}, params, &critical_stats);
+
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.56),
+                       Catch::Matchers::WithinAbs(0.00, 0.01));
+            CHECK_THAT(std::abs(ct_cc - 1.46), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+        }
+        SECTION("Kinks are not allowed")
+        {
+            params.operational_params.op_condition = is_operational_params::operational_condition::REJECT_KINKS;
+            const auto ct_qe =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_fan_out_tt()}, params, &critical_stats);
+
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.56),
+                       Catch::Matchers::WithinAbs(0.00, 0.01));
+            CHECK_THAT(std::abs(ct_qe - 1.46), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+            params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+            const auto ct_cc =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_fan_out_tt()}, params, &critical_stats);
+
+            CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.56),
+                       Catch::Matchers::WithinAbs(0.00, 0.01));
+            CHECK_THAT(std::abs(ct_cc - 1.46), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+        }
     }
 
-    SECTION("OR gate")
+    SECTION("Bestagon CX gate")
+    {
+        const auto crossing_lyt = blueprints::bestagon_crossing<TestType>();
+
+        params.operational_params.simulation_parameters = sim_params;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
+
+        const auto ct_qe = critical_temperature_gate_based(crossing_lyt, std::vector<tt>{create_crossing_wire_tt()},
+                                                           params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.32),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_qe - 0.84999999999999998), Catch::Matchers::WithinAbs(0.000000, 0.000001));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc = critical_temperature_gate_based(crossing_lyt, std::vector<tt>{create_crossing_wire_tt()},
+                                                           params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.32),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_cc - 0.84999999999999998), Catch::Matchers::WithinAbs(0.000000, 0.000001));
+
+#endif  // FICTION_ALGLIB_ENABLED
+    }
+
+    SECTION("SiQAD OR gate")
     {
         lyt.assign_cell_type({0, 0, 0}, sidb_technology::cell_type::INPUT);
         lyt.assign_cell_type({26, 0, 0}, sidb_technology::cell_type::INPUT);
@@ -338,16 +526,50 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
         sim_params.mu_minus = -0.25;
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_or_tt()}, params, &critical_stats);
+        SECTION("Kinks are allowed")
+        {
+            const auto ct_qe =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_or_tt()}, params, &critical_stats);
 
-        CHECK(ct < 350);
+            CHECK(ct_qe < 350);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+            params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+            const auto ct_cc =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_or_tt()}, params, &critical_stats);
+
+            CHECK(ct_cc < 350);
+
+#endif  // FICTION_ALGLIB_ENABLED
+        }
+        SECTION("Kinks are not allowed")
+        {
+            params.operational_params.op_condition = is_operational_params::operational_condition::REJECT_KINKS;
+            const auto ct_qe =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_or_tt()}, params, &critical_stats);
+
+            CHECK(ct_qe < 350);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+            params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+            const auto ct_cc =
+                critical_temperature_gate_based(lyt, std::vector<tt>{create_or_tt()}, params, &critical_stats);
+
+            CHECK(ct_cc < 350);
+
+#endif  // FICTION_ALGLIB_ENABLED
+        }
     }
 
     SECTION("Not working diagonal Wire")
@@ -371,20 +593,36 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
 
         lyt.assign_cell_type({36, 19, 0}, sidb_technology::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
 
         CHECK(critical_stats.algorithm_name == "QuickExact");
 
         CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
                    Catch::Matchers::WithinAbs(305.95, 0.01));
-        CHECK_THAT(std::abs(ct), Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_qe), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
+
+        CHECK(critical_stats.algorithm_name == "ClusterComplete");
+
+        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                   Catch::Matchers::WithinAbs(305.95, 0.01));
+        CHECK_THAT(std::abs(ct_cc), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("nine SiDBs, QuickSim, non-gate-based")
@@ -400,23 +638,35 @@ TEMPLATE_TEST_CASE("Test critical_temperature function", "[critical-temperature]
         lyt.assign_cell_type({9, 1, 1}, sidb_technology::cell_type::NORMAL);
         lyt.assign_cell_type({12, 1, 1}, sidb_technology::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::APPROXIMATE;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 750;
-        params.iteration_steps       = 500;
-        params.alpha                 = 0.6;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKSIM;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 750;
+        params.iteration_steps                          = 500;
+        params.alpha                                    = 0.6;
 
-        const auto ct = critical_temperature_non_gate_based(lyt, params, &critical_stats);
+        const auto ct_qe = critical_temperature_non_gate_based(lyt, params, &critical_stats);
 
         CHECK(critical_stats.algorithm_name == "QuickSim");
 
-        CHECK_THAT(std::abs(ct), Catch::Matchers::WithinAbs(11.55, 0.01));
+        CHECK_THAT(std::abs(ct_qe), Catch::Matchers::WithinAbs(11.55, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc = critical_temperature_non_gate_based(lyt, params, &critical_stats);
+
+        CHECK(critical_stats.algorithm_name == "ClusterComplete");
+
+        CHECK_THAT(std::abs(ct_cc), Catch::Matchers::WithinAbs(11.55, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 }
 
 TEMPLATE_TEST_CASE("Test critical_temperature function, using offset coordinates", "[critical-temperature]",
-                   sidb_100_cell_clk_lyt, cds_sidb_100_cell_clk_lyt)
+                   sidb_100_cell_clk_lyt)
 {
     TestType lyt{};
 
@@ -427,35 +677,35 @@ TEMPLATE_TEST_CASE("Test critical_temperature function, using offset coordinates
 
     SECTION("No physically valid charge distribution could be found")
     {
-        lyt.assign_cell_type({0, 0, 0}, sidb_technology::cell_type::INPUT);
+        lyt.assign_cell_type({0, 0}, sidb_technology::cell_type::INPUT);
         lyt.assign_cell_type({2, 2}, sidb_technology::cell_type::INPUT);
         lyt.assign_cell_type({6, 2}, sidb_technology::cell_type::OUTPUT);
         lyt.assign_cell_type({8, 2}, sidb_technology::cell_type::OUTPUT);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::APPROXIMATE;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 0;
-        params.alpha                 = 0.0;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKSIM;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 0;
+        params.alpha                                    = 0.0;
 
-        const auto ct =
+        const auto ct_qs =
             critical_temperature_gate_based<TestType>(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
 
         CHECK(critical_stats.algorithm_name == "QuickSim");
 
         CHECK(critical_stats.num_valid_lyt == 0);
-        CHECK(ct == 0.0);
+        CHECK(ct_qs == 0.0);
     }
 
     SECTION("One SiDB")
     {
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
         const auto ct = critical_temperature_gate_based<TestType>(lyt, std::vector<tt>{tt{}}, params, &critical_stats);
 
@@ -488,16 +738,28 @@ TEMPLATE_TEST_CASE("Test critical_temperature function, using offset coordinates
 
         lyt.assign_cell_type({36, 38, 0}, sidb_technology::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
 
-        CHECK(ct == 0.0);
+        CHECK(ct_qe == 0.0);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_id_tt()}, params, &critical_stats);
+
+        CHECK(ct_cc == 0.0);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("four SiDBs with two valid charge distributions, QuickExact")
@@ -508,21 +770,36 @@ TEMPLATE_TEST_CASE("Test critical_temperature function, using offset coordinates
         lyt.assign_cell_type({2, 0}, TestType::cell_type::NORMAL);
         lyt.assign_cell_type({2, 4}, TestType::cell_type::NORMAL);
 
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_non_gate_based(lyt, params, &critical_stats);
+        const auto ct_qe = critical_temperature_non_gate_based(lyt, params, &critical_stats);
 
         CHECK(critical_stats.algorithm_name == "QuickExact");
 
         CHECK(critical_stats.num_valid_lyt == 2);
         CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
                    Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
-        CHECK(ct == 350);
+        CHECK(ct_qe == 350);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc = critical_temperature_non_gate_based(lyt, params, &critical_stats);
+
+        CHECK(critical_stats.algorithm_name == "ClusterComplete");
+
+        CHECK(critical_stats.num_valid_lyt == 2);
+        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                   Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
+        CHECK(ct_cc == 350);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 
     SECTION("Y-shape SiDB AND gate")
@@ -544,24 +821,42 @@ TEMPLATE_TEST_CASE("Test critical_temperature function, using offset coordinates
 
         lyt.assign_cell_type({10, 19}, sidb_technology::cell_type::NORMAL);
 
-        sim_params.mu_minus          = -0.28;
-        params.simulation_parameters = sim_params;
-        params.engine                = critical_temperature_params::simulation_engine::EXACT;
-        params.confidence_level      = 0.99;
-        params.max_temperature       = 350;
-        params.iteration_steps       = 80;
-        params.alpha                 = 0.7;
+        sim_params.mu_minus                             = -0.28;
+        params.operational_params.simulation_parameters = sim_params;
+        params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+        params.confidence_level                         = 0.99;
+        params.max_temperature                          = 350;
+        params.iteration_steps                          = 80;
+        params.alpha                                    = 0.7;
 
-        const auto ct = critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
 
         CHECK(critical_stats.algorithm_name == "QuickExact");
 
         CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
                    Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
-        CHECK(ct == 350);
+        CHECK(ct_qe == 350);
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt, std::vector<tt>{create_and_tt()}, params, &critical_stats);
+
+        CHECK(critical_stats.algorithm_name == "ClusterComplete");
+
+        CHECK_THAT(std::abs(critical_stats.energy_between_ground_state_and_first_erroneous),
+                   Catch::Matchers::WithinAbs(std::numeric_limits<double>::infinity(), 0.01));
+        CHECK(ct_cc == 350);
+
+#endif  // FICTION_ALGLIB_ENABLED
     }
 }
 
+// to save runtime in the CI, this test is only run in RELEASE mode
+#ifdef NDEBUG
 TEMPLATE_TEST_CASE("Critical temperature of Bestagon CX, QuickExact", "[critical-temperature], [quality]",
                    sidb_100_cell_clk_lyt_siqad, cds_sidb_100_cell_clk_lyt_siqad)
 {
@@ -607,21 +902,33 @@ TEMPLATE_TEST_CASE("Critical temperature of Bestagon CX, QuickExact", "[critical
     lyt.assign_cell_type({2, 19, 0}, sidb_technology::cell_type::NORMAL);
     lyt.assign_cell_type({36, 19, 0}, sidb_technology::cell_type::NORMAL);
 
-    critical_temperature_params params{};
-    sidb_simulation_parameters  sim_params{2, -0.32, 5.6, 5.0};
+    critical_temperature_params      params{};
+    const sidb_simulation_parameters sim_params{2, -0.32, 5.6, 5.0};
 
     critical_temperature_stats critical_stats{};
 
-    params.simulation_parameters = sim_params;
-    params.engine                = critical_temperature_params::simulation_engine::EXACT;
-    params.confidence_level      = 0.99;
-    params.max_temperature       = 350;
+    params.operational_params.simulation_parameters = sim_params;
+    params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+    params.confidence_level                         = 0.99;
+    params.max_temperature                          = 350;
 
-    const auto ct = critical_temperature_gate_based(lyt, create_crossing_wire_tt(), params, &critical_stats);
+    const auto ct_qe = critical_temperature_gate_based(lyt, create_crossing_wire_tt(), params, &critical_stats);
 
     CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.32),
                Catch::Matchers::WithinAbs(0.00, 0.01));
-    CHECK_THAT(std::abs(ct - 0.85), Catch::Matchers::WithinAbs(0.00, 0.01));
+    CHECK_THAT(std::abs(ct_qe - 0.85), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+    params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+    const auto ct_cc = critical_temperature_gate_based(lyt, create_crossing_wire_tt(), params, &critical_stats);
+
+    CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.32),
+               Catch::Matchers::WithinAbs(0.00, 0.01));
+    CHECK_THAT(std::abs(ct_cc - 0.85), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
 }
 
 TEMPLATE_TEST_CASE("Critical temperature of Bestagon double wire, QuickExact", "[critical-temperature], [quality]",
@@ -634,17 +941,56 @@ TEMPLATE_TEST_CASE("Critical temperature of Bestagon double wire, QuickExact", "
 
     critical_temperature_stats critical_stats{};
 
-    params.simulation_parameters = sim_params;
-    params.engine                = critical_temperature_params::simulation_engine::EXACT;
-    params.confidence_level      = 0.99;
-    params.max_temperature       = 350;
+    params.operational_params.simulation_parameters = sim_params;
+    params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+    params.confidence_level                         = 0.99;
+    params.max_temperature                          = 350;
 
-    const auto ct =
-        critical_temperature_gate_based(lyt_double_wire_gate, create_double_wire_tt(), params, &critical_stats);
+    SECTION("Kinks are allowed")
+    {
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt_double_wire_gate, create_double_wire_tt(), params, &critical_stats);
 
-    CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 10.717),
-               Catch::Matchers::WithinAbs(0.00, 0.01));
-    CHECK_THAT(std::abs(ct - 24.18), Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 10.717),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_qe - 24.18), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt_double_wire_gate, create_double_wire_tt(), params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 10.717),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_cc - 24.18), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+    }
+    SECTION("Kinks are not allowed")
+    {
+        params.operational_params.op_condition = is_operational_params::operational_condition::REJECT_KINKS;
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt_double_wire_gate, create_double_wire_tt(), params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 10.717),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_qe - 23.77), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt_double_wire_gate, create_double_wire_tt(), params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 10.717),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_cc - 23.77), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+    }
 }
 
 TEMPLATE_TEST_CASE("Critical temperature of Bestagon half adder gate, QuickExact", "[critical-temperature], [quality]",
@@ -657,15 +1003,55 @@ TEMPLATE_TEST_CASE("Critical temperature of Bestagon half adder gate, QuickExact
 
     critical_temperature_stats critical_stats{};
 
-    params.simulation_parameters = sim_params;
-    params.engine                = critical_temperature_params::simulation_engine::EXACT;
-    params.confidence_level      = 0.99;
-    params.max_temperature       = 350;
+    params.operational_params.simulation_parameters = sim_params;
+    params.operational_params.sim_engine            = sidb_simulation_engine::QUICKEXACT;
+    params.confidence_level                         = 0.99;
+    params.max_temperature                          = 350;
 
-    const auto ct =
-        critical_temperature_gate_based(lyt_half_adder_gate, create_half_adder_tt(), params, &critical_stats);
+    SECTION("Kinks are allowed")
+    {
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt_half_adder_gate, create_half_adder_tt(), params, &critical_stats);
 
-    CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.15),
-               Catch::Matchers::WithinAbs(0.00, 0.01));
-    CHECK_THAT(std::abs(ct - 0.40), Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.15),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_qe - 0.40), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt_half_adder_gate, create_half_adder_tt(), params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.15),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_cc - 0.40), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+    }
+    SECTION("Kinks are not allowed")
+    {
+        params.operational_params.op_condition = is_operational_params::operational_condition::REJECT_KINKS;
+        const auto ct_qe =
+            critical_temperature_gate_based(lyt_half_adder_gate, create_half_adder_tt(), params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.15),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_qe - 0.40), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#if (FICTION_ALGLIB_ENABLED)
+
+        params.operational_params.sim_engine = sidb_simulation_engine::CLUSTERCOMPLETE;
+
+        const auto ct_cc =
+            critical_temperature_gate_based(lyt_half_adder_gate, create_half_adder_tt(), params, &critical_stats);
+
+        CHECK_THAT(std::fabs(critical_stats.energy_between_ground_state_and_first_erroneous - 0.15),
+                   Catch::Matchers::WithinAbs(0.00, 0.01));
+        CHECK_THAT(std::abs(ct_cc - 0.40), Catch::Matchers::WithinAbs(0.00, 0.01));
+
+#endif  // FICTION_ALGLIB_ENABLED
+    }
 }
+#endif
