@@ -10,7 +10,11 @@
 #include <fiction/algorithms/simulation/sidb/detect_bdl_wires.hpp>
 #include <fiction/algorithms/simulation/sidb/is_operational.hpp>
 #include <fiction/algorithms/simulation/sidb/sidb_simulation_engine.hpp>
+#include <fiction/algorithms/simulation/sidb/sidb_simulation_parameters.hpp>
+#include <fiction/layouts/cell_level_layout.hpp>
 #include <fiction/technology/cell_technologies.hpp>
+#include <fiction/technology/sidb_defects.hpp>
+#include <fiction/traits.hpp>
 #include <fiction/types.hpp>
 #include <fiction/utils/truth_table_utils.hpp>
 
@@ -23,52 +27,121 @@ using namespace fiction;
 
 TEST_CASE("SiQAD OR gate", "[is-operational]")
 {
-    const auto layout_or_gate = blueprints::siqad_or_gate<sidb_cell_clk_lyt_siqad>();
+    const auto or_gate = blueprints::siqad_or_gate<sidb_cell_clk_lyt_siqad>();
 
-    const sidb_100_cell_clk_lyt_siqad lat{layout_or_gate};
+    const sidb_100_cell_clk_lyt_siqad lat{or_gate};
 
-    SECTION("Reject kinks and keep simulation results")
+    auto op_params = is_operational_params{
+        sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT,
+        bdl_input_iterator_params{detect_bdl_wires_params{1.5},
+                                  bdl_input_iterator_params::input_bdl_configuration::PERTURBER_ABSENCE_ENCODED},
+        is_operational_params::operational_condition::TOLERATE_KINKS, {}, is_operational_params::simulation_results_mode::KEEP_SIMULATION_RESULTS};
+
+    SECTION("determine if layout is operational, tolerate kinks and keep simulation results")
     {
-
-        const auto p_no = is_operational(
-            lat, std::vector<tt>{create_or_tt()},
-            is_operational_params{sidb_simulation_parameters{2, -0.28}, sidb_simulation_engine::QUICKEXACT,
-                                  bdl_input_iterator_params{
-                                      detect_bdl_wires_params{1.5},
-                                      bdl_input_iterator_params::input_bdl_configuration::PERTURBER_ABSENCE_ENCODED},
-                                  operational_condition::REJECT_KINKS,
-                                  is_operational_params::simulation_results_mode::KEEP_SIMULATION_RESULTS});
-        CHECK(p_no.first == operational_status::NON_OPERATIONAL);
-        CHECK(!p_no.second.simulation_results.has_value());
+        const auto [op_status, aux_stats] = is_operational(lat, std::vector<tt>{create_or_tt()}, op_params);
+        CHECK(op_status == operational_status::OPERATIONAL);
+        CHECK(aux_stats.simulation_results.has_value());
     }
 
-    SECTION("Tolerate kinks and keep simulation results")
+    // from now on, we will discard simulation results
+    op_params.simulation_results_retention = is_operational_params::simulation_results_mode::DISCARD_SIMULATION_RESULTS;
+
+    SECTION("determine if layout is operational, tolerate kinks and discard simulation results")
     {
-        const auto p_o = is_operational(
-            lat, std::vector<tt>{create_or_tt()},
-            is_operational_params{sidb_simulation_parameters{2, -0.28}, sidb_simulation_engine::QUICKEXACT,
-                                  bdl_input_iterator_params{
-                                      detect_bdl_wires_params{1.5},
-                                      bdl_input_iterator_params::input_bdl_configuration::PERTURBER_ABSENCE_ENCODED},
-                                  operational_condition::TOLERATE_KINKS,
-                                  is_operational_params::simulation_results_mode::KEEP_SIMULATION_RESULTS});
-        CHECK(p_o.first == operational_status::OPERATIONAL);
-        REQUIRE(p_o.second.simulation_results.has_value());
-        CHECK(p_o.second.simulation_results.value().size() == 4);
+        const auto [op_status, aux_stats] = is_operational(lat, std::vector<tt>{create_or_tt()}, op_params);
+        CHECK(op_status == operational_status::OPERATIONAL);
+        CHECK(!aux_stats.simulation_results.has_value());
     }
 
-    SECTION("Tolerate kinks and discard simulation results")
+    // from now on, we will reject kinks
+    op_params.op_condition = is_operational_params::operational_condition::REJECT_KINKS;
+
+    SECTION("determine if layout is operational, reject kinks")
     {
-        const auto p_o2 = is_operational(
-            lat, std::vector<tt>{create_or_tt()},
-            is_operational_params{sidb_simulation_parameters{2, -0.28}, sidb_simulation_engine::QUICKEXACT,
-                                  bdl_input_iterator_params{
-                                      detect_bdl_wires_params{1.5},
-                                      bdl_input_iterator_params::input_bdl_configuration::PERTURBER_ABSENCE_ENCODED},
-                                  operational_condition::TOLERATE_KINKS,
-                                  is_operational_params::simulation_results_mode::DISCARD_SIMULATION_RESULTS});
-        CHECK(p_o2.first == operational_status::OPERATIONAL);
-        CHECK(!p_o2.second.simulation_results.has_value());
+        CHECK(is_operational(lat, std::vector<tt>{create_and_tt()}, op_params).first ==
+              operational_status::NON_OPERATIONAL);
+    }
+
+    SECTION("determine if kinks induce layout to become non-operational")
+    {
+        const auto kink_induced_non_operational =
+            is_kink_induced_non_operational(lat, std::vector<tt>{create_or_tt()}, op_params);
+        CHECK(kink_induced_non_operational);
+    }
+
+    const auto input_wires  = detect_bdl_wires(lat, detect_bdl_wires_params{1.5}, bdl_wire_selection::INPUT);
+    const auto output_wires = detect_bdl_wires(lat, detect_bdl_wires_params{1.5}, bdl_wire_selection::OUTPUT);
+
+    REQUIRE(input_wires.size() == 2);
+
+    CHECK(input_wires[0].pairs.size() == 2);
+    CHECK(input_wires[1].pairs.size() == 2);
+
+    CHECK(output_wires.size() == 1);
+
+    SECTION("use pre-determined I/O pins")
+    {
+        CHECK(is_operational(lat, std::vector<tt>{create_and_tt()}, op_params, input_wires, output_wires).first ==
+              operational_status::NON_OPERATIONAL);
+    }
+
+    SECTION("determine if kinks induce layout to become non-operational")
+    {
+        CHECK(is_kink_induced_non_operational(lat, std::vector<tt>{create_or_tt()}, op_params, input_wires,
+                                              output_wires));
+    }
+
+    SECTION("determine input patterns for which kinks induce layout to become non-operational")
+    {
+        const auto kink_induced_non_operational_input_pattern =
+            kink_induced_non_operational_input_patterns(lat, std::vector<tt>{create_or_tt()}, op_params);
+
+        CHECK(kink_induced_non_operational_input_pattern.size() == 1);
+
+        op_params.op_condition = is_operational_params::operational_condition::TOLERATE_KINKS;
+        CHECK(is_operational(lat, std::vector<tt>{create_or_tt()}, op_params).first == operational_status::OPERATIONAL);
+    }
+}
+
+TEST_CASE("SiQAD NAND gate", "[is-operational]")
+{
+    const auto nand_gate = blueprints::siqad_nand_gate<sidb_cell_clk_lyt_siqad>();
+
+    const sidb_100_cell_clk_lyt_siqad lat{nand_gate};
+
+    auto op_params = is_operational_params{
+        sidb_simulation_parameters{2, -0.28}, sidb_simulation_engine::QUICKEXACT,
+        bdl_input_iterator_params{detect_bdl_wires_params{1.5},
+                                  bdl_input_iterator_params::input_bdl_configuration::PERTURBER_ABSENCE_ENCODED},
+        is_operational_params::operational_condition::REJECT_KINKS,
+        is_operational_params::operational_analysis_strategy::FILTER_THEN_SIMULATION};
+
+    SECTION("Pruning and simulation")
+    {
+        CHECK(is_operational(lat, std::vector<tt>{create_nand_tt()}, op_params).first ==
+              operational_status::OPERATIONAL);
+    }
+    SECTION("only pruning")
+    {
+        op_params.strategy_to_analyze_operational_status =
+            is_operational_params::operational_analysis_strategy::FILTER_ONLY;
+        CHECK(is_operational(lat, std::vector<tt>{create_nand_tt()}, op_params).first ==
+              operational_status::OPERATIONAL);
+    }
+
+    const auto input_wires  = detect_bdl_wires(lat, detect_bdl_wires_params{2.0}, bdl_wire_selection::INPUT);
+    const auto output_wires = detect_bdl_wires(lat, detect_bdl_wires_params{2.0}, bdl_wire_selection::OUTPUT);
+
+    sidb_100_cell_clk_lyt_siqad canvas_lyt{};
+    canvas_lyt.assign_cell_type({10, 4, 1}, sidb_technology::cell_type::NORMAL);
+    canvas_lyt.assign_cell_type({10, 5, 1}, sidb_technology::cell_type::NORMAL);
+
+    SECTION("use pre-determined I/O pins")
+    {
+        CHECK(is_operational(lat, std::vector<tt>{create_nand_tt()}, op_params, input_wires, output_wires,
+                             std::optional{canvas_lyt})
+                  .first == operational_status::OPERATIONAL);
     }
 }
 
@@ -132,6 +205,22 @@ TEST_CASE("Bestagon FO2 gate", "[is-operational]")
                   is_operational_params{sidb_simulation_parameters{2, -0.30}, sidb_simulation_engine::QUICKSIM})
                   .first == operational_status::NON_OPERATIONAL);
     }
+
+#if (FICTION_ALGLIB_ENABLED)
+
+    SECTION("using ClusterComplete")
+    {
+        CHECK(is_operational(
+                  lyt, std::vector<tt>{create_fan_out_tt()},
+                  is_operational_params{sidb_simulation_parameters{3, -0.32}, sidb_simulation_engine::CLUSTERCOMPLETE})
+                  .first == operational_status::OPERATIONAL);
+        CHECK(is_operational(
+                  lyt, std::vector<tt>{create_fan_out_tt()},
+                  is_operational_params{sidb_simulation_parameters{3, -0.30}, sidb_simulation_engine::CLUSTERCOMPLETE})
+                  .first == operational_status::NON_OPERATIONAL);
+    }
+
+#endif  // FICTION_ALGLIB_ENABLED
 }
 
 TEST_CASE("Bestagon CROSSING gate", "[is-operational]")
@@ -154,10 +243,43 @@ TEST_CASE("Bestagon CROSSING gate", "[is-operational]")
 
 TEST_CASE("Bestagon AND gate", "[is-operational]")
 {
-    const auto lyt = blueprints::bestagon_and<sidb_cell_clk_lyt_siqad>();
+    auto lyt = blueprints::bestagon_and<sidb_defect_cell_clk_lyt_siqad>();
 
-    CHECK(lyt.num_cells() == 23);
+    const sidb_simulation_parameters params{2, -0.32};
 
+    SECTION("Without defects")
+    {
+        CHECK(lyt.num_cells() == 23);
+
+        CHECK(is_operational(
+                  lyt, std::vector<tt>{create_and_tt()},
+                  is_operational_params{sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT})
+                  .first == operational_status::OPERATIONAL);
+        CHECK(is_operational(
+                  lyt, std::vector<tt>{create_and_tt()},
+                  is_operational_params{sidb_simulation_parameters{2, -0.30}, sidb_simulation_engine::QUICKEXACT})
+                  .first == operational_status::NON_OPERATIONAL);
+    }
+    SECTION("With defects")
+    {
+        lyt.assign_sidb_defect({3, 16, 1},
+                               sidb_defect{sidb_defect_type::UNKNOWN, -1, params.epsilon_r, params.lambda_tf});
+        CHECK(is_operational(lyt, std::vector<tt>{create_and_tt()},
+                             is_operational_params{params, sidb_simulation_engine::QUICKEXACT})
+                  .first == operational_status::OPERATIONAL);
+
+        // move defect one to the right
+        lyt.move_sidb_defect({3, 16, 1}, {4, 16, 1});
+        CHECK(is_operational(lyt, std::vector<tt>{create_and_tt()},
+                             is_operational_params{params, sidb_simulation_engine::QUICKEXACT})
+                  .first == operational_status::OPERATIONAL);
+
+        // move defect one to the right
+        lyt.move_sidb_defect({4, 16, 1}, {5, 16, 1});
+        CHECK(is_operational(lyt, std::vector<tt>{create_and_tt()},
+                             is_operational_params{params, sidb_simulation_engine::QUICKEXACT})
+                  .first == operational_status::NON_OPERATIONAL);
+    }
     SECTION("Check operation for different values of mu")
     {
         CHECK(is_operational(
@@ -169,7 +291,7 @@ TEST_CASE("Bestagon AND gate", "[is-operational]")
                   is_operational_params{sidb_simulation_parameters{2, -0.30}, sidb_simulation_engine::QUICKEXACT})
                   .first == operational_status::NON_OPERATIONAL);
     }
-    SECTION("Count the number of non-operational input combinations")
+    SECTION("Count the number of non-operational input combinations, accepting kinks")
     {
         const auto op_inputs = operational_input_patterns(
             lyt, std::vector<tt>{create_and_tt()},
@@ -217,7 +339,7 @@ TEMPLATE_TEST_CASE("AND gate on the H-Si(111)-1x1 surface", "[is-operational]", 
 {
     const auto lyt = blueprints::and_gate_111<TestType>();
 
-    SECTION("Check operation for different values of mu")
+    SECTION("check operation for different values of mu")
     {
         const auto op_inputs = operational_input_patterns(
             lyt, std::vector<tt>{create_and_tt()},
@@ -225,7 +347,7 @@ TEMPLATE_TEST_CASE("AND gate on the H-Si(111)-1x1 surface", "[is-operational]", 
         CHECK(op_inputs.size() == 4);
         CHECK(op_inputs == std::set<uint64_t>{0, 1, 2, 3});
     }
-    SECTION("Count the number of non-operational input combinations")
+    SECTION("count the number of non-operational input combinations")
     {
         const auto op_inputs = operational_input_patterns(
             lyt, std::vector<tt>{create_and_tt()},
@@ -234,7 +356,7 @@ TEMPLATE_TEST_CASE("AND gate on the H-Si(111)-1x1 surface", "[is-operational]", 
         CHECK(op_inputs == std::set<uint64_t>{0, 3});
     }
 
-    SECTION("Verify the operational status of the AND gate, which is mirrored on the x-axis. Note that the input BDL "
+    SECTION("verify the operational status of the AND gate, which is mirrored on the x-axis. Note that the input BDL "
             "pairs are located at the bottom, while the output BDL pairs are at the top.")
     {
         const auto lyt_mirrored_x = blueprints::and_gate_111_mirrored_on_the_x_axis<TestType>();
@@ -246,7 +368,9 @@ TEMPLATE_TEST_CASE("AND gate on the H-Si(111)-1x1 surface", "[is-operational]", 
     }
 }
 
-TEST_CASE("AND gate with bestagon structure and kink state at right input wire for input 01", "[is-operational]")
+TEST_CASE(
+    "AND gate with Bestagon structure and kink state on right input wire for input 01 and left input wire for input 10",
+    "[is-operational]")
 {
     const auto lyt = blueprints::and_gate_with_kink_states<sidb_cell_clk_lyt_siqad>();
 
@@ -256,71 +380,151 @@ TEST_CASE("AND gate with bestagon structure and kink state at right input wire f
                              is_operational_params{sidb_simulation_parameters{2, -0.32}})
                   .first == operational_status::OPERATIONAL);
     }
-    SECTION("Forbid kink states")
+    SECTION("reject kink states")
     {
         CHECK(is_operational(lyt, std::vector<tt>{create_and_tt()},
                              is_operational_params{sidb_simulation_parameters{2, -0.32},
                                                    sidb_simulation_engine::QUICKEXACT, bdl_input_iterator_params{},
-                                                   operational_condition::REJECT_KINKS})
+                                                   is_operational_params::operational_condition::REJECT_KINKS})
                   .first == operational_status::NON_OPERATIONAL);
+    }
+    SECTION("check if is_kink_induced_non_operational returns true")
+    {
+        // check if the function works correctly even if the parameter is wrong (kinks are accepted).
+        CHECK(is_kink_induced_non_operational(
+            lyt, std::vector<tt>{create_and_tt()},
+            is_operational_params{sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT,
+                                  bdl_input_iterator_params{},
+                                  is_operational_params::operational_condition::TOLERATE_KINKS}));
+    }
+
+    SECTION("check input patterns for which kinks induce the layout to become non-operational")
+    {
+        CHECK(kink_induced_non_operational_input_patterns(
+                  lyt, std::vector<tt>{create_and_tt()},
+                  is_operational_params{sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT,
+                                        bdl_input_iterator_params{},
+                                        is_operational_params::operational_condition::TOLERATE_KINKS}) ==
+              std::set<uint64_t>{1, 2});
     }
 }
 
+TEST_CASE("BDL wire", "[is-operational]")
+{
+    using layout = sidb_cell_clk_lyt_siqad;
+
+    layout lyt{{24, 0}, "BDL wire"};
+
+    lyt.assign_cell_type({0, 0, 0}, sidb_technology::cell_type::INPUT);
+    lyt.assign_cell_type({3, 0, 0}, sidb_technology::cell_type::INPUT);
+
+    lyt.assign_cell_type({6, 0, 0}, sidb_technology::cell_type::NORMAL);
+    lyt.assign_cell_type({8, 0, 0}, sidb_technology::cell_type::NORMAL);
+
+    lyt.assign_cell_type({12, 0, 0}, sidb_technology::cell_type::NORMAL);
+    lyt.assign_cell_type({14, 0, 0}, sidb_technology::cell_type::NORMAL);
+
+    lyt.assign_cell_type({18, 0, 0}, sidb_technology::cell_type::OUTPUT);
+    lyt.assign_cell_type({20, 0, 0}, sidb_technology::cell_type::OUTPUT);
+
+    // output perturber
+    lyt.assign_cell_type({24, 0, 0}, sidb_technology::cell_type::NORMAL);
+
+    const sidb_100_cell_clk_lyt_siqad lat{lyt};
+
+    sidb_simulation_parameters sim_params{};
+
+    sim_params.base = 2;
+
+    const is_operational_params params{sim_params};
+
+    CHECK(is_operational(lyt, std::vector<tt>{create_id_tt()}, params).first == operational_status::OPERATIONAL);
+}
+
+TEST_CASE("Special wire that cannot be pruned, but is non-operational when kinks are rejected", "[is-operational]")
+{
+    sidb_cell_clk_lyt_siqad lyt{};
+
+    // input wires
+    lyt.assign_cell_type({0, 0, 0}, sidb_cell_clk_lyt_siqad::cell_type::INPUT);
+    lyt.assign_cell_type({2, 1, 0}, sidb_cell_clk_lyt_siqad::cell_type::INPUT);
+
+    lyt.assign_cell_type({6, 2, 0}, sidb_cell_clk_lyt_siqad::cell_type::NORMAL);
+    lyt.assign_cell_type({8, 3, 0}, sidb_cell_clk_lyt_siqad::cell_type::NORMAL);
+
+    lyt.assign_cell_type({14, 5, 0}, sidb_cell_clk_lyt_siqad::cell_type::NORMAL);
+    lyt.assign_cell_type({12, 4, 0}, sidb_cell_clk_lyt_siqad::cell_type::NORMAL);
+
+    // canvas SiDBs
+    lyt.assign_cell_type({11, 7, 0}, sidb_cell_clk_lyt_siqad::cell_type::LOGIC);
+    lyt.assign_cell_type({13, 13, 0}, sidb_cell_clk_lyt_siqad::cell_type::LOGIC);
+
+    // output wires
+    lyt.assign_cell_type({14, 15, 0}, sidb_cell_clk_lyt_siqad::cell_type::NORMAL);
+    lyt.assign_cell_type({12, 16, 0}, sidb_cell_clk_lyt_siqad::cell_type::NORMAL);
+
+    lyt.assign_cell_type({8, 17, 0}, sidb_cell_clk_lyt_siqad::cell_type::OUTPUT);
+    lyt.assign_cell_type({6, 18, 0}, sidb_cell_clk_lyt_siqad::cell_type::OUTPUT);
+
+    lyt.assign_cell_type({2, 19, 0}, sidb_cell_clk_lyt_siqad::cell_type::NORMAL);
+
+    sidb_simulation_parameters sim_params{};
+
+    sim_params.base = 2;
+
+    is_operational_params params{sim_params};
+
+    SECTION("Rejecting Kinks")
+    {
+        params.op_condition = is_operational_params::operational_condition::REJECT_KINKS;
+        params.strategy_to_analyze_operational_status =
+            is_operational_params::operational_analysis_strategy::FILTER_THEN_SIMULATION;
+
+        CHECK(is_operational(lyt, std::vector<tt>{create_id_tt()}, params).first ==
+              operational_status::NON_OPERATIONAL);
+    }
+
+    SECTION("Only conducting pruning and tolerating kinks")
+    {
+        params.op_condition = is_operational_params::operational_condition::TOLERATE_KINKS;
+        params.strategy_to_analyze_operational_status =
+            is_operational_params::operational_analysis_strategy::FILTER_ONLY;
+
+        CHECK(is_operational(lyt, std::vector<tt>{create_id_tt()}, params).first ==
+              operational_status::NON_OPERATIONAL);
+    }
+}
+
+// to save runtime in the CI, this test is only run in RELEASE mode
+#ifdef NDEBUG
 TEST_CASE("flipped CX bestagon gate", "[is-operational]")
 {
     const auto lyt = blueprints::crossing_bestagon_shape_input_down_output_up<sidb_cell_clk_lyt_siqad>();
 
     CHECK(is_operational(lyt, create_crossing_wire_tt(),
                          is_operational_params{sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT,
-                                               bdl_input_iterator_params{}, operational_condition::REJECT_KINKS})
+                                               bdl_input_iterator_params{},
+                                               is_operational_params::operational_condition::REJECT_KINKS})
               .first == operational_status::OPERATIONAL);
+
+    const auto kink_induced_non_operational_input_pattern = kink_induced_non_operational_input_patterns(
+        lyt, create_crossing_wire_tt(),
+        is_operational_params{sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT,
+                              bdl_input_iterator_params{}, is_operational_params::operational_condition::REJECT_KINKS});
+
+    CHECK(kink_induced_non_operational_input_pattern.empty());
+
+    const auto kink_induced_non_operational = is_kink_induced_non_operational(
+        lyt, create_crossing_wire_tt(),
+        is_operational_params{sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT,
+                              bdl_input_iterator_params{}, is_operational_params::operational_condition::REJECT_KINKS});
+
+    CHECK(!kink_induced_non_operational);
 }
 
 TEST_CASE("is operational check for Bestagon CX gate", "[is-operational], [quality]")
 {
-    using layout = sidb_cell_clk_lyt_siqad;
-
-    layout lyt{};
-
-    lyt.assign_cell_type({36, 1, 0}, sidb_technology::cell_type::INPUT);
-    lyt.assign_cell_type({2, 1, 0}, sidb_technology::cell_type::INPUT);
-
-    lyt.assign_cell_type({0, 0, 0}, sidb_technology::cell_type::INPUT);
-    lyt.assign_cell_type({38, 0, 0}, sidb_technology::cell_type::INPUT);
-
-    lyt.assign_cell_type({6, 2, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({20, 12, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({8, 3, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({14, 5, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({14, 11, 1}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({12, 4, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({14, 15, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({26, 4, 0}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({14, 9, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({24, 15, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({12, 16, 0}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({18, 9, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({26, 16, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({24, 13, 1}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({24, 5, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({30, 3, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({16, 13, 1}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({32, 2, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({20, 8, 0}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({30, 17, 0}, sidb_technology::cell_type::OUTPUT);
-    lyt.assign_cell_type({6, 18, 0}, sidb_technology::cell_type::OUTPUT);
-
-    lyt.assign_cell_type({32, 18, 0}, sidb_technology::cell_type::OUTPUT);
-    lyt.assign_cell_type({8, 17, 0}, sidb_technology::cell_type::OUTPUT);
-
-    lyt.assign_cell_type({2, 19, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({36, 19, 0}, sidb_technology::cell_type::NORMAL);
+    const auto lyt = blueprints::bestagon_crossing<sidb_cell_clk_lyt_siqad>();
 
     CHECK(lyt.num_cells() == 29);
 
@@ -346,13 +550,30 @@ TEST_CASE("is operational check for Bestagon CX gate", "[is-operational], [quali
         CHECK(is_operational(
                   lat, create_crossing_wire_tt(),
                   is_operational_params{sidb_simulation_parameters{2, -0.32}, sidb_simulation_engine::QUICKEXACT},
-                  std::optional{input_bdl_wires}, std::optional{output_bdl_wires})
+                  input_bdl_wires, output_bdl_wires)
                   .first == operational_status::OPERATIONAL);
         CHECK(is_operational(
                   lat, create_crossing_wire_tt(),
                   is_operational_params{sidb_simulation_parameters{2, -0.30}, sidb_simulation_engine::QUICKEXACT},
-                  std::optional{input_bdl_wires}, std::optional{output_bdl_wires})
+                  input_bdl_wires, output_bdl_wires)
                   .first == operational_status::NON_OPERATIONAL);
+        CHECK(!is_kink_induced_non_operational(
+            lat, create_crossing_wire_tt(),
+            is_operational_params{sidb_simulation_parameters{2, -0.30}, sidb_simulation_engine::QUICKEXACT},
+            input_bdl_wires, output_bdl_wires));
+    }
+
+    SECTION("using predetermined wires and only applying pruning without simulation")
+    {
+        const auto input_bdl_wires  = detect_bdl_wires(lat, detect_bdl_wires_params{}, bdl_wire_selection::INPUT);
+        const auto output_bdl_wires = detect_bdl_wires(lat, detect_bdl_wires_params{}, bdl_wire_selection::OUTPUT);
+
+        auto op_params = is_operational_params{sidb_simulation_parameters{2, -0.32}};
+        op_params.strategy_to_analyze_operational_status =
+            is_operational_params::operational_analysis_strategy::FILTER_ONLY;
+
+        CHECK(is_operational(lat, create_crossing_wire_tt(), op_params, input_bdl_wires, output_bdl_wires).first ==
+              operational_status::OPERATIONAL);
     }
 }
 
@@ -391,35 +612,4 @@ TEST_CASE("is operational check for Bestagon half adder", "[is-operational], [qu
                        is_operational_params{sidb_simulation_parameters{2, -0.25}, sidb_simulation_engine::QUICKEXACT})
             .first == operational_status::NON_OPERATIONAL);
 }
-
-TEST_CASE("BDL wire", "[is-operational]")
-{
-    using layout = sidb_cell_clk_lyt_siqad;
-
-    layout lyt{{24, 0}, "BDL wire"};
-
-    lyt.assign_cell_type({0, 0, 0}, sidb_technology::cell_type::INPUT);
-    lyt.assign_cell_type({3, 0, 0}, sidb_technology::cell_type::INPUT);
-
-    lyt.assign_cell_type({6, 0, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({8, 0, 0}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({12, 0, 0}, sidb_technology::cell_type::NORMAL);
-    lyt.assign_cell_type({14, 0, 0}, sidb_technology::cell_type::NORMAL);
-
-    lyt.assign_cell_type({18, 0, 0}, sidb_technology::cell_type::OUTPUT);
-    lyt.assign_cell_type({20, 0, 0}, sidb_technology::cell_type::OUTPUT);
-
-    // output perturber
-    lyt.assign_cell_type({24, 0, 0}, sidb_technology::cell_type::NORMAL);
-
-    const sidb_100_cell_clk_lyt_siqad lat{lyt};
-
-    sidb_simulation_parameters sim_params{};
-
-    sim_params.base = 2;
-
-    const is_operational_params params{sim_params};
-
-    CHECK(is_operational(lyt, std::vector<tt>{create_id_tt()}, params).first == operational_status::OPERATIONAL);
-}
+#endif
