@@ -7,8 +7,8 @@
 
 #include "utils/blueprints/layout_blueprints.hpp"
 
-#include <fiction/algorithms/simulation/sidb/determine_groundstate_from_simulation_results.hpp>
 #include <fiction/algorithms/simulation/sidb/energy_distribution.hpp>
+#include <fiction/algorithms/simulation/sidb/groundstate_from_simulation_result.hpp>
 #include <fiction/algorithms/simulation/sidb/quicksim.hpp>
 #include <fiction/algorithms/simulation/sidb/sidb_simulation_result.hpp>
 #include <fiction/layouts/coordinates.hpp>
@@ -1002,7 +1002,7 @@ TEMPLATE_TEST_CASE("QuickSim gate simulation on the Si-111 surface", "[quicksim]
 
     const auto simulation_results = quicksim<TestType>(lyt, params);
 
-    const auto ground_state = determine_groundstate_from_simulation_results(simulation_results);
+    const auto ground_state = groundstate_from_simulation_result(simulation_results);
     REQUIRE(ground_state.size() == 1);
 
     CHECK(ground_state.front().get_charge_state({0, 0, 0}) == sidb_charge_state::NEGATIVE);
@@ -1026,7 +1026,7 @@ TEMPLATE_TEST_CASE("QuickSim AND gate simulation on the Si-111 surface", "[quick
 
         const auto simulation_results = quicksim<TestType>(lyt, params);
 
-        const auto ground_state = determine_groundstate_from_simulation_results(simulation_results);
+        const auto ground_state = groundstate_from_simulation_result(simulation_results);
         REQUIRE(ground_state.size() == 1);
 
         CHECK(ground_state.front().get_charge_state({0, 0, 0}) == sidb_charge_state::NEGATIVE);
@@ -1066,7 +1066,7 @@ TEMPLATE_TEST_CASE("QuickSim AND gate simulation on the Si-111 surface", "[quick
 
         const auto simulation_results = quicksim<TestType>(lyt, params);
 
-        const auto ground_state = determine_groundstate_from_simulation_results(simulation_results);
+        const auto ground_state = groundstate_from_simulation_result(simulation_results);
 
         REQUIRE(ground_state.size() == 1);
 
