@@ -1777,6 +1777,86 @@ TEST_CASE("4-phase Ripple", "[clocking-scheme]")
     CHECK(ripple4({3 + 4, 3 + 4}) == 0);
 }
 
+TEST_CASE("4-phase SRS", "[clocking-scheme]")
+{
+    using clk_lyt = clocked_layout<cartesian_layout<offset::ucoord_t>>;
+
+    const auto srs4 = srs_clocking<clk_lyt>();
+
+    CHECK(srs4.num_clocks == 4u);
+    CHECK(srs4.max_in_degree == 3u);
+    CHECK(srs4.max_out_degree == 3u);
+    CHECK(srs4.is_regular());
+
+    CHECK(srs4({0, 0}) == 1);
+    CHECK(srs4({0, 1}) == 2);
+    CHECK(srs4({0, 2}) == 3);
+    CHECK(srs4({0, 3}) == 0);
+    CHECK(srs4({1, 0}) == 0);
+    CHECK(srs4({1, 1}) == 3);
+    CHECK(srs4({1, 2}) == 2);
+    CHECK(srs4({1, 3}) == 1);
+    CHECK(srs4({2, 0}) == 3);
+    CHECK(srs4({2, 1}) == 0);
+    CHECK(srs4({2, 2}) == 3);
+    CHECK(srs4({2, 3}) == 2);
+    CHECK(srs4({3, 0}) == 2);
+    CHECK(srs4({3, 1}) == 1);
+    CHECK(srs4({3, 2}) == 0);
+    CHECK(srs4({3, 3}) == 1);
+
+    CHECK(srs4({0 + 4, 0}) == 1);
+    CHECK(srs4({0 + 4, 1}) == 2);
+    CHECK(srs4({0 + 4, 2}) == 3);
+    CHECK(srs4({0 + 4, 3}) == 0);
+    CHECK(srs4({1 + 4, 0}) == 0);
+    CHECK(srs4({1 + 4, 1}) == 3);
+    CHECK(srs4({1 + 4, 2}) == 2);
+    CHECK(srs4({1 + 4, 3}) == 1);
+    CHECK(srs4({2 + 4, 0}) == 3);
+    CHECK(srs4({2 + 4, 1}) == 0);
+    CHECK(srs4({2 + 4, 2}) == 3);
+    CHECK(srs4({2 + 4, 3}) == 2);
+    CHECK(srs4({3 + 4, 0}) == 2);
+    CHECK(srs4({3 + 4, 1}) == 1);
+    CHECK(srs4({3 + 4, 2}) == 0);
+    CHECK(srs4({3 + 4, 3}) == 1);
+
+    CHECK(srs4({0, 0 + 4}) == 1);
+    CHECK(srs4({0, 1 + 4}) == 2);
+    CHECK(srs4({0, 2 + 4}) == 3);
+    CHECK(srs4({0, 3 + 4}) == 0);
+    CHECK(srs4({1, 0 + 4}) == 0);
+    CHECK(srs4({1, 1 + 4}) == 3);
+    CHECK(srs4({1, 2 + 4}) == 2);
+    CHECK(srs4({1, 3 + 4}) == 1);
+    CHECK(srs4({2, 0 + 4}) == 3);
+    CHECK(srs4({2, 1 + 4}) == 0);
+    CHECK(srs4({2, 2 + 4}) == 3);
+    CHECK(srs4({2, 3 + 4}) == 2);
+    CHECK(srs4({3, 0 + 4}) == 2);
+    CHECK(srs4({3, 1 + 4}) == 1);
+    CHECK(srs4({3, 2 + 4}) == 0);
+    CHECK(srs4({3, 3 + 4}) == 1);
+
+    CHECK(srs4({0 + 4, 0 + 4}) == 1);
+    CHECK(srs4({0 + 4, 1 + 4}) == 2);
+    CHECK(srs4({0 + 4, 2 + 4}) == 3);
+    CHECK(srs4({0 + 4, 3 + 4}) == 0);
+    CHECK(srs4({1 + 4, 0 + 4}) == 0);
+    CHECK(srs4({1 + 4, 1 + 4}) == 3);
+    CHECK(srs4({1 + 4, 2 + 4}) == 2);
+    CHECK(srs4({1 + 4, 3 + 4}) == 1);
+    CHECK(srs4({2 + 4, 0 + 4}) == 3);
+    CHECK(srs4({2 + 4, 1 + 4}) == 0);
+    CHECK(srs4({2 + 4, 2 + 4}) == 3);
+    CHECK(srs4({2 + 4, 3 + 4}) == 2);
+    CHECK(srs4({3 + 4, 0 + 4}) == 2);
+    CHECK(srs4({3 + 4, 1 + 4}) == 1);
+    CHECK(srs4({3 + 4, 2 + 4}) == 0);
+    CHECK(srs4({3 + 4, 3 + 4}) == 1);
+}
+
 TEST_CASE("3-phase BANCS", "[clocking-scheme]")
 {
     using clk_lyt = clocked_layout<cartesian_layout<offset::ucoord_t>>;
@@ -1905,84 +1985,84 @@ TEST_CASE("Override clocking", "[clocking-scheme]")
     CHECK(twoddwave4({3, 3}) == 2);
 }
 
-TEST_CASE("4-phase ESP", "[clocking-scheme]")
+TEST_CASE("4-phase ESR", "[clocking-scheme]")
 {
     using clk_lyt = clocked_layout<cartesian_layout<offset::ucoord_t>>;
 
-    const auto esp4 = esr_clocking<clk_lyt>();
+    const auto esr4 = esr_clocking<clk_lyt>();
 
-    CHECK(esp4.num_clocks == 4u);
-    CHECK(esp4.max_in_degree == 3u);
-    CHECK(esp4.max_out_degree == 3u);
-    CHECK(esp4.is_regular());
+    CHECK(esr4.num_clocks == 4u);
+    CHECK(esr4.max_in_degree == 3u);
+    CHECK(esr4.max_out_degree == 3u);
+    CHECK(esr4.is_regular());
 
-    CHECK(esp4({0, 0}) == 3);
-    CHECK(esp4({0, 1}) == 0);
-    CHECK(esp4({0, 2}) == 1);
-    CHECK(esp4({0, 3}) == 0);
-    CHECK(esp4({1, 0}) == 0);
-    CHECK(esp4({1, 1}) == 1);
-    CHECK(esp4({1, 2}) == 2);
-    CHECK(esp4({1, 3}) == 3);
-    CHECK(esp4({2, 0}) == 1);
-    CHECK(esp4({2, 1}) == 2);
-    CHECK(esp4({2, 2}) == 3);
-    CHECK(esp4({2, 3}) == 2);
-    CHECK(esp4({3, 0}) == 2);
-    CHECK(esp4({3, 1}) == 3);
-    CHECK(esp4({3, 2}) == 0);
-    CHECK(esp4({3, 3}) == 1);
+    CHECK(esr4({0, 0}) == 3);
+    CHECK(esr4({0, 1}) == 0);
+    CHECK(esr4({0, 2}) == 1);
+    CHECK(esr4({0, 3}) == 0);
+    CHECK(esr4({1, 0}) == 0);
+    CHECK(esr4({1, 1}) == 1);
+    CHECK(esr4({1, 2}) == 2);
+    CHECK(esr4({1, 3}) == 3);
+    CHECK(esr4({2, 0}) == 1);
+    CHECK(esr4({2, 1}) == 2);
+    CHECK(esr4({2, 2}) == 3);
+    CHECK(esr4({2, 3}) == 2);
+    CHECK(esr4({3, 0}) == 2);
+    CHECK(esr4({3, 1}) == 3);
+    CHECK(esr4({3, 2}) == 0);
+    CHECK(esr4({3, 3}) == 1);
 
-    CHECK(esp4({0 + 4, 0}) == 3);
-    CHECK(esp4({0 + 4, 1}) == 0);
-    CHECK(esp4({0 + 4, 2}) == 1);
-    CHECK(esp4({0 + 4, 3}) == 0);
-    CHECK(esp4({1 + 4, 0}) == 0);
-    CHECK(esp4({1 + 4, 1}) == 1);
-    CHECK(esp4({1 + 4, 2}) == 2);
-    CHECK(esp4({1 + 4, 3}) == 3);
-    CHECK(esp4({2 + 4, 0}) == 1);
-    CHECK(esp4({2 + 4, 1}) == 2);
-    CHECK(esp4({2 + 4, 2}) == 3);
-    CHECK(esp4({2 + 4, 3}) == 2);
-    CHECK(esp4({3 + 4, 0}) == 2);
-    CHECK(esp4({3 + 4, 1}) == 3);
-    CHECK(esp4({3 + 4, 2}) == 0);
-    CHECK(esp4({3 + 4, 3}) == 1);
+    CHECK(esr4({0 + 4, 0}) == 3);
+    CHECK(esr4({0 + 4, 1}) == 0);
+    CHECK(esr4({0 + 4, 2}) == 1);
+    CHECK(esr4({0 + 4, 3}) == 0);
+    CHECK(esr4({1 + 4, 0}) == 0);
+    CHECK(esr4({1 + 4, 1}) == 1);
+    CHECK(esr4({1 + 4, 2}) == 2);
+    CHECK(esr4({1 + 4, 3}) == 3);
+    CHECK(esr4({2 + 4, 0}) == 1);
+    CHECK(esr4({2 + 4, 1}) == 2);
+    CHECK(esr4({2 + 4, 2}) == 3);
+    CHECK(esr4({2 + 4, 3}) == 2);
+    CHECK(esr4({3 + 4, 0}) == 2);
+    CHECK(esr4({3 + 4, 1}) == 3);
+    CHECK(esr4({3 + 4, 2}) == 0);
+    CHECK(esr4({3 + 4, 3}) == 1);
 
-    CHECK(esp4({0, 0 + 4}) == 3);
-    CHECK(esp4({0, 1 + 4}) == 0);
-    CHECK(esp4({0, 2 + 4}) == 1);
-    CHECK(esp4({0, 3 + 4}) == 0);
-    CHECK(esp4({1, 0 + 4}) == 0);
-    CHECK(esp4({1, 1 + 4}) == 1);
-    CHECK(esp4({1, 2 + 4}) == 2);
-    CHECK(esp4({1, 3 + 4}) == 3);
-    CHECK(esp4({2, 0 + 4}) == 1);
-    CHECK(esp4({2, 1 + 4}) == 2);
-    CHECK(esp4({2, 2 + 4}) == 3);
-    CHECK(esp4({2, 3 + 4}) == 2);
-    CHECK(esp4({3, 0 + 4}) == 2);
-    CHECK(esp4({3, 1 + 4}) == 3);
-    CHECK(esp4({3, 2 + 4}) == 0);
-    CHECK(esp4({3, 3 + 4}) == 1);
+    CHECK(esr4({0, 0 + 4}) == 3);
+    CHECK(esr4({0, 1 + 4}) == 0);
+    CHECK(esr4({0, 2 + 4}) == 1);
+    CHECK(esr4({0, 3 + 4}) == 0);
+    CHECK(esr4({1, 0 + 4}) == 0);
+    CHECK(esr4({1, 1 + 4}) == 1);
+    CHECK(esr4({1, 2 + 4}) == 2);
+    CHECK(esr4({1, 3 + 4}) == 3);
+    CHECK(esr4({2, 0 + 4}) == 1);
+    CHECK(esr4({2, 1 + 4}) == 2);
+    CHECK(esr4({2, 2 + 4}) == 3);
+    CHECK(esr4({2, 3 + 4}) == 2);
+    CHECK(esr4({3, 0 + 4}) == 2);
+    CHECK(esr4({3, 1 + 4}) == 3);
+    CHECK(esr4({3, 2 + 4}) == 0);
+    CHECK(esr4({3, 3 + 4}) == 1);
 
-    CHECK(esp4({0 + 4, 0 + 4}) == 3);
-    CHECK(esp4({0 + 4, 1 + 4}) == 0);
-    CHECK(esp4({0 + 4, 2 + 4}) == 1);
-    CHECK(esp4({0 + 4, 3 + 4}) == 0);
-    CHECK(esp4({1 + 4, 0 + 4}) == 0);
-    CHECK(esp4({1 + 4, 1 + 4}) == 1);
-    CHECK(esp4({1 + 4, 2 + 4}) == 2);
-    CHECK(esp4({1 + 4, 3 + 4}) == 3);
-    CHECK(esp4({2 + 4, 0 + 4}) == 1);
-    CHECK(esp4({2 + 4, 1 + 4}) == 2);
-    CHECK(esp4({2 + 4, 2 + 4}) == 3);
-    CHECK(esp4({2 + 4, 3 + 4}) == 2);
-    CHECK(esp4({3 + 4, 0 + 4}) == 2);
-    CHECK(esp4({3 + 4, 1 + 4}) == 3);
-    CHECK(esp4({3 + 4, 2 + 4}) == 0);
-    CHECK(esp4({3 + 4, 3 + 4}) == 1);
+    CHECK(esr4({0 + 4, 0 + 4}) == 3);
+    CHECK(esr4({0 + 4, 1 + 4}) == 0);
+    CHECK(esr4({0 + 4, 2 + 4}) == 1);
+    CHECK(esr4({0 + 4, 3 + 4}) == 0);
+    CHECK(esr4({1 + 4, 0 + 4}) == 0);
+    CHECK(esr4({1 + 4, 1 + 4}) == 1);
+    CHECK(esr4({1 + 4, 2 + 4}) == 2);
+    CHECK(esr4({1 + 4, 3 + 4}) == 3);
+    CHECK(esr4({2 + 4, 0 + 4}) == 1);
+    CHECK(esr4({2 + 4, 1 + 4}) == 2);
+    CHECK(esr4({2 + 4, 2 + 4}) == 3);
+    CHECK(esr4({2 + 4, 3 + 4}) == 2);
+    CHECK(esr4({3 + 4, 0 + 4}) == 2);
+    CHECK(esr4({3 + 4, 1 + 4}) == 3);
+    CHECK(esr4({3 + 4, 2 + 4}) == 0);
+    CHECK(esr4({3 + 4, 3 + 4}) == 1);
 }
 
 TEST_CASE("Clocking lookup", "[clocking-scheme]")
@@ -2009,6 +2089,7 @@ TEST_CASE("Clocking lookup", "[clocking-scheme]")
     check({"esr", "ESR", "eSr", "EsR"}, clock_name::ESR);
     check({"cfe", "CFE", "cFe", "CfE"}, clock_name::CFE);
     check({"ripple", "RIPPLE", "RiPpLe", "RIppLE"}, clock_name::RIPPLE);
+    check({"srs", "SRS", "sRs", "SrS"}, clock_name::SRS);
     check({"bancs", "BANCS", "BaNCs", "banCS"}, clock_name::BANCS);
 
     CHECK(!get_clocking_scheme<clk_lyt>("").has_value());
@@ -2021,6 +2102,7 @@ TEST_CASE("Clocking lookup", "[clocking-scheme]")
     CHECK(!get_clocking_scheme<clk_lyt>("ERS").has_value());
     CHECK(!get_clocking_scheme<clk_lyt>("CEF").has_value());
     CHECK(!get_clocking_scheme<clk_lyt>("RPIPLE").has_value());
+    CHECK(!get_clocking_scheme<clk_lyt>("SSR").has_value());
     CHECK(!get_clocking_scheme<clk_lyt>("BNCS").has_value());
 }
 
@@ -2028,16 +2110,26 @@ TEST_CASE("Linear schemes", "[clocking-scheme]")
 {
     using clk_lyt = clocked_layout<cartesian_layout<offset::ucoord_t>>;
 
-    CHECK(is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::COLUMNAR)));
-    CHECK(is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::ROW)));
-    CHECK(is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::TWODDWAVE)));
-    CHECK(is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::TWODDWAVE_HEX)));
+    auto check_linear_scheme = [](const auto& name, bool expected)
+    {
+        auto cs = get_clocking_scheme<clk_lyt>(name);
+        REQUIRE(cs.has_value());
+        CHECK(is_linear_scheme<clk_lyt>(*cs) == expected);
+    };
 
-    CHECK(!is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::OPEN)));
-    CHECK(!is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::USE)));
-    CHECK(!is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::RES)));
-    CHECK(!is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::ESR)));
-    CHECK(!is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::CFE)));
-    CHECK(!is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::RIPPLE)));
-    CHECK(!is_linear_scheme<clk_lyt>(*get_clocking_scheme<clk_lyt>(clock_name::BANCS)));
+    // Linear clocking schemes
+    check_linear_scheme(clock_name::COLUMNAR, true);
+    check_linear_scheme(clock_name::ROW, true);
+    check_linear_scheme(clock_name::TWODDWAVE, true);
+    check_linear_scheme(clock_name::TWODDWAVE_HEX, true);
+
+    // Non-linear clocking schemes
+    check_linear_scheme(clock_name::OPEN, false);
+    check_linear_scheme(clock_name::USE, false);
+    check_linear_scheme(clock_name::RES, false);
+    check_linear_scheme(clock_name::ESR, false);
+    check_linear_scheme(clock_name::CFE, false);
+    check_linear_scheme(clock_name::RIPPLE, false);
+    check_linear_scheme(clock_name::SRS, false);
+    check_linear_scheme(clock_name::BANCS, false);
 }
